Pre-generate target types
Add missing purchase options targets

Add new components

Fix output type

Update components

Fix non-rendering target components
