+++ packages/ui-extensions/loom.config.ts
@@ -1,10 +1,18 @@
 import {createPackage} from '@shopify/loom';
-import {existsSync, readFileSync, writeFileSync} from 'fs';
+import {
+  existsSync,
+  readFileSync,
+  mkdirSync,
+  writeFileSync,
+  readdirSync,
+  copyFileSync,
+} from 'fs';
 import {join, resolve} from 'path';
 
 import {defaultProjectPlugin} from '../../config/loom';
 import {rollupPlugins} from '@shopify/loom-plugin-build-library';
 import replace from '@rollup/plugin-replace';
+import {Project} from 'ts-morph';
 
 const packageJSON = JSON.parse(
   readFileSync(resolve(__dirname, './package.json')).toString(),
@@ -32,22 +40,73 @@ export default createPackage((pkg) => {
       {
         name: 'add-components-types',
         closeBundle: async () => {
-          const mainTypesPath = join(__dirname, 'build/ts/surfaces/admin.d.ts');
-          const componentsTypes = join(
+          const buildPath = join(
             __dirname,
-            'src/surfaces/admin/components.d.ts',
+            'build/ts/surfaces/admin/components',
           );
-          if (existsSync(mainTypesPath) && existsSync(componentsTypes)) {
-            const mainTypesContent = readFileSync(mainTypesPath).toString();
-            const componentsTypesContent = readFileSync(componentsTypes)
-              .toString()
-              .replaceAll(/\/\*.*$/g, '');
+          const srcPath = join(__dirname, 'src/surfaces/admin/components');
+          if (existsSync(srcPath) && existsSync(buildPath)) {
+            const components = readdirSync(srcPath);
 
-            writeFileSync(
-              mainTypesPath,
-              componentsTypesContent.concat(mainTypesContent),
-            );
+            components.forEach((componentFile) => {
+              if (componentFile.endsWith('d.ts')) {
+                copyFileSync(
+                  join(srcPath, componentFile),
+                  join(buildPath, componentFile),
+                );
+              }
+            });
           }
+        },
+      },
+      {
+        name: 'add-target-types',
+        closeBundle: async () => {
+          // See https://stackoverflow.com/questions/43909566/get-keys-of-a-typescript-interface-as-array-of-strings#answer-62477194
+          const project = new Project();
+          const sourceFile = project.addSourceFileAtPath(
+            resolve(__dirname, `./src/surfaces/admin/extension-targets.ts`),
+          );
+          const node = sourceFile.getInterface('ExtensionTargets')!;
+          const targets = node.getProperties().map((property) => {
+            const componentSet = property
+              .getType()
+              .getProperty('components')
+              ?.getTypeAtLocation(node)
+              .getText();
+
+            return {target: property.getName(), componentSet};
+          });
+
+          targets.forEach(({target, componentSet}) => {
+            const parts = target.replaceAll("'", '').split('.');
+            const surface = parts[0];
+            const fileName = `${parts.join('.')}.d.ts`;
+            const directory = resolve(
+              __dirname,
+              join(`./build/ts/surfaces/${surface}/targets`),
+            );
+            const targetPath = join(directory, fileName);
+            const componentsReference = componentSet
+              ? `/// <reference types="../components/${componentSet?.replaceAll(
+                  '"',
+                  '',
+                )}" />\n`
+              : '';
+            const template = `import type {ExtensionTargets} from '../extension-targets';
+type Target = ExtensionTargets[${target}];
+export type Api = Target['api'];
+export type Output = Target['output'];
+
+declare global {
+  const shopify: Api;
+  export default function extension(): Output;
+}\n`;
+            if (!existsSync(directory)) {
+              mkdirSync(directory);
+            }
+            writeFileSync(targetPath, componentsReference.concat(template));
+          });
         },
       },
     ]),
