+++ packages/ui-extensions/package.json
@@ -15,6 +15,9 @@
     "*": {
       "admin": [
         "./build/ts/surfaces/admin.d.ts"
+      ],
+      "admin.*": [
+        "./build/ts/surfaces/admin/targets/admin.*"
       ]
     }
   },
@@ -30,6 +33,9 @@
       "esnext": "./build/esnext/surfaces/admin.esnext",
       "import": "./build/esm/surfaces/admin.mjs",
       "require": "./build/cjs/surfaces/admin.js"
+    },
+    "./admin.*": {
+      "types": "./build/ts/surfaces/admin/targets/admin.*.d.ts"
     },
     "./preact": {
       "types": "./build/ts/preact.d.ts",
@@ -65,5 +71,8 @@
     "rules": {
       "@shopify/strict-component-boundaries": "off"
     }
+  },
+  "dependencies": {
+    "ts-morph": "^25.0.1"
   }
 }
