+++ packages/ui-extensions/src/surfaces/admin.ts
@@ -1,4 +1,2 @@
 export * from './admin/api';
 export * from './admin/extension-targets';
-export * from './admin/extension';
-export * from './admin/globals';
