+++ packages/ui-extensions/src/surfaces/admin/api.ts
@@ -9,3 +9,4 @@ export type {ProductDetailsConfigurationApi} from './api/product-configuration/p
 export type {ProductVariantDetailsConfigurationApi} from './api/product-configuration/product-variant-details-configuration';
 export type {OrderRoutingRuleApi} from './api/order-routing-rule/order-routing-rule';
 export type {ValidationSettingsApi} from './api/checkout-rules/validation-settings';
+export type {PurchaseOptionsCardConfigurationApi} from './api/purchase-options-card-action';
