+++ packages/ui-extensions/src/surfaces/admin/api/purchase-options-card-action.ts
@@ -1,0 +1,10 @@
+import {ActionExtensionApi} from './action/action';
+import type {ExtensionTarget as AnyExtensionTarget} from '../extension-targets';
+
+export interface PurchaseOptionsCardConfigurationApi<
+  ExtensionTarget extends AnyExtensionTarget,
+> extends ActionExtensionApi<ExtensionTarget> {
+  data: {
+    selected: {id: string; sellingPlanId?: string}[];
+  };
+}
