+++ packages/ui-extensions/src/surfaces/admin/components/ActionExtensionComponents.d.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
+/* eslint-disable spaced-comment */
+/// <reference types="./StandardComponents"  />
+/// <reference types="./AdminAction.d.ts" />
