+++ packages/ui-extensions/src/surfaces/admin/components/BlockExtensionComponents.d.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
+/* eslint-disable spaced-comment */
+/// <reference types="./StandardComponents" />
+/// <reference types="./AdminBlock.d.ts" />
