+++ packages/ui-extensions/src/surfaces/admin/components/Select.d.ts
@@ -1,0 +1,161 @@
+/** VERSION: 0.45.0 **/
+/* eslint-disable import/extensions */
+
+/* eslint-disable @typescript-eslint/no-namespace */
+/* eslint-disable @typescript-eslint/member-ordering */
+
+import type {SelectProps$1, ComponentChild} from './shared.d.ts';
+
+export type Styles = string;
+export type RenderImpl = Omit<ShadowRootInit, 'mode'> & {
+  ShadowRoot: (element: any) => ComponentChild;
+  styles?: Styles;
+};
+export interface ActivationEventEsque {
+  shiftKey: boolean;
+  metaKey: boolean;
+  ctrlKey: boolean;
+  button: number;
+}
+export interface ClickOptions {
+  /**
+   * The event you want to influence the synthetic click.
+   */
+  sourceEvent?: ActivationEventEsque;
+}
+/**
+ * Base class for creating custom elements with Preact.
+ * While this class could be used in both Node and the browser, the constructor will only be used in the browser.
+ * So we give it a type of HTMLElement to avoid typing issues later where it's used, which will only happen in the browser.
+ */
+declare const BaseClass: typeof globalThis.HTMLElement;
+declare abstract class PreactCustomElement extends BaseClass {
+  /** @private */
+  static get observedAttributes(): string[];
+  constructor({
+    styles,
+    ShadowRoot: renderFunction,
+    delegatesFocus,
+    ...options
+  }: RenderImpl);
+
+  /** @private */
+  attributeChangedCallback(name: string): void;
+  /** @private */
+  connectedCallback(): void;
+  /** @private */
+  disconnectedCallback(): void;
+  /** @private */
+  adoptedCallback(): void;
+  /**
+   * Queue a run of the render function.
+   * You shouldn't need to call this manually - it should be handled by changes to @property values.
+   * @private
+   */
+  queueRender(): void;
+  /**
+   * Like the standard `element.click()`, but you can influence the behavior with a `sourceEvent`.
+   *
+   * For example, if the `sourceEvent` was a middle click, or has particular keys held down,
+   * components will attempt to produce the desired behavior on links, such as opening the page in the background tab.
+   * @private
+   * @param options
+   */
+  click({sourceEvent}?: ClickOptions): void;
+}
+
+export type CallbackEvent<T extends keyof HTMLElementTagNameMap> = Event & {
+  target: HTMLElementTagNameMap[T];
+};
+export type CallbackEventListener<T extends keyof HTMLElementTagNameMap> =
+  | (EventListener & {
+      (event: CallbackEvent<T>): void;
+    })
+  | null;
+
+declare const internals: unique symbol;
+export type PreactInputProps = Required<
+  Pick<TextFieldProps, 'disabled' | 'id' | 'name' | 'value'>
+>;
+declare class PreactInputElement
+  extends PreactCustomElement
+  implements PreactInputProps
+{
+  static formAssociated: boolean;
+  /** @private */
+  [internals]: ElementInternals;
+  protected getDefaultValue(): string;
+  accessor onchange: CallbackEventListener<'input'>;
+  accessor oninput: CallbackEventListener<'input'>;
+  accessor disabled: PreactInputProps['disabled'];
+  accessor id: PreactInputProps['id'];
+  accessor name: PreactInputProps['name'];
+  get value(): PreactInputProps['value'];
+  set value(value: PreactInputProps['value']);
+  constructor(renderImpl: RenderImpl);
+}
+
+export type SelectProps = PreactInputProps &
+  Required<
+    Pick<
+      SelectProps$1,
+      | 'details'
+      | 'disabled'
+      | 'error'
+      | 'label'
+      | 'name'
+      | 'placeholder'
+      | 'required'
+      | 'value'
+      | 'icon'
+      | 'labelAccessibilityVisibility'
+    >
+  >;
+
+declare const tagName = 's-select';
+export interface ReactProps extends Partial<SelectProps> {
+  onChange?: (event: CallbackEvent<typeof tagName>) => void;
+  onInput?: (event: CallbackEvent<typeof tagName>) => void;
+  onBlur?: (event: CallbackEvent<typeof tagName>) => void;
+  onFocus?: (event: CallbackEvent<typeof tagName>) => void;
+}
+
+declare const usedFirstOptionSymbol: unique symbol;
+declare const hasInitialValueSymbol: unique symbol;
+
+declare class Select extends PreactInputElement implements SelectProps {
+  accessor icon: SelectProps['icon'];
+  accessor details: SelectProps['details'];
+  accessor error: SelectProps['error'];
+  accessor label: SelectProps['label'];
+  accessor placeholder: SelectProps['placeholder'];
+  accessor required: SelectProps['required'];
+  accessor labelAccessibilityVisibility: SelectProps['labelAccessibilityVisibility'];
+  connectedCallback(): void;
+  disconnectedCallback(): void;
+  constructor();
+  /**
+   * used to determine if no value or defaultValue was set, in which case the first non-disabled option was used
+   *
+   * this is important because we need to use the placeholder in these situations, even though the first value will be submitted as part of the form
+   */
+  [usedFirstOptionSymbol]: boolean;
+  [hasInitialValueSymbol]: boolean;
+  get value(): string;
+  set value(value: string);
+  formResetCallback(): void;
+}
+declare global {
+  interface HTMLElementTagNameMap {
+    [tagName]: Select;
+  }
+}
+declare module 'preact' {
+  namespace createElement.JSX {
+    interface IntrinsicElements {
+      [tagName]: HTMLAttributes<HTMLElement> & ReactProps;
+    }
+  }
+}
+
+export {Select};
