+++ packages/ui-extensions/src/surfaces/admin/components/StandardComponents.d.ts
@@ -1,0 +1,40 @@
+/* eslint-disable spaced-comment */
+/* eslint-disable @typescript-eslint/triple-slash-reference */
+/// <reference types="./Badge.d.ts"  />
+/// <reference types="./Banner.d.ts"/>
+/// <reference types="./Box.d.ts"/>
+/// <reference types="./Button.d.ts"/>
+/// <reference types="./Checkbox.d.ts"/>
+/// <reference types="./Choice.d.ts"/>
+/// <reference types="./Choicelist.d.ts"/>
+/// <reference types="./Clickable.d.ts"/>
+/// <reference types="./Divider.d.ts"/>
+/// <reference types="./EmailField.d.ts"/>
+/// <reference types="./Grid.d.ts"/>
+/// <reference types="./Heading.d.ts"/>
+/// <reference types="./Icon.d.ts"/>
+/// <reference types="./Image.d.ts"/>
+/// <reference types="./Link.d.ts"/>
+/// <reference types="./ListItem.d.ts"/>
+/// <reference types="./MoneyField.d.ts"/>
+/// <reference types="./NumberField.d.ts"/>
+/// <reference types="./Option.d.ts"/>
+/// <reference types="./OptionGroup.d.ts"/>
+/// <reference types="./OrderedList.d.ts"/>
+/// <reference types="./Paragraph.d.ts"/>
+/// <reference types="./PasswordField.d.ts"/>
+/// <reference types="./Section.d.ts"/>
+/// <reference types="./Select.d.ts"/>
+/// <reference types="./Spinner.d.ts"/>
+/// <reference types="./Stack.d.ts"/>
+/// <reference types="./Table.d.ts"/>
+/// <reference types="./TableBody.d.ts"/>
+/// <reference types="./TableCell.d.ts"/>
+/// <reference types="./TableHeader.d.ts"/>
+/// <reference types="./TableHeaderRow.d.ts"/>
+/// <reference types="./TableRow.d.ts"/>
+/// <reference types="./Text.d.ts"/>
+/// <reference types="./TextArea.d.ts"/>
+/// <reference types="./TextField.d.ts"/>
+/// <reference types="./UnorderedList.d.ts"/>
+/// <reference types="./URLField.d.ts"/>
