+++ packages/ui-extensions/src/surfaces/admin/components/Text.d.ts
@@ -1,0 +1,114 @@
+/** VERSION: 0.45.0 **/
+/* eslint-disable import/extensions */
+
+/* eslint-disable @typescript-eslint/no-namespace */
+/* eslint-disable @typescript-eslint/member-ordering */
+
+import type {TextProps$1, ComponentChild} from './shared.d.ts';
+
+export interface TextProps
+  extends Required<
+    Pick<TextProps$1, 'display' | 'accessibilityVisibility' | 'dir'>
+  > {
+  color: Extract<TextProps$1['color'], 'base' | 'subdued'>;
+  type: Extract<
+    TextProps$1['type'],
+    'address' | 'redundant' | 'strong' | 'generic'
+  >;
+  tone: Extract<
+    TextProps$1['tone'],
+    'auto' | 'info' | 'success' | 'warning' | 'critical' | 'neutral' | 'caution'
+  >;
+  fontVariantNumeric: Extract<
+    TextProps$1['fontVariantNumeric'],
+    'auto' | 'normal' | 'tabular-nums'
+  >;
+}
+
+declare const tagName = 's-text';
+export interface ReactProps
+  extends Partial<TextProps>,
+    Pick<TextProps$1, 'id'> {}
+
+export type Styles = string;
+export type RenderImpl = Omit<ShadowRootInit, 'mode'> & {
+  ShadowRoot: (element: any) => ComponentChild;
+  styles?: Styles;
+};
+export interface ActivationEventEsque {
+  shiftKey: boolean;
+  metaKey: boolean;
+  ctrlKey: boolean;
+  button: number;
+}
+export interface ClickOptions {
+  /**
+   * The event you want to influence the synthetic click.
+   */
+  sourceEvent?: ActivationEventEsque;
+}
+/**
+ * Base class for creating custom elements with Preact.
+ * While this class could be used in both Node and the browser, the constructor will only be used in the browser.
+ * So we give it a type of HTMLElement to avoid typing issues later where it's used, which will only happen in the browser.
+ */
+declare const BaseClass: typeof globalThis.HTMLElement;
+declare abstract class PreactCustomElement extends BaseClass {
+  /** @private */
+  static get observedAttributes(): string[];
+  constructor({
+    styles,
+    ShadowRoot: renderFunction,
+    delegatesFocus,
+    ...options
+  }: RenderImpl);
+
+  /** @private */
+  attributeChangedCallback(name: string): void;
+  /** @private */
+  connectedCallback(): void;
+  /** @private */
+  disconnectedCallback(): void;
+  /** @private */
+  adoptedCallback(): void;
+  /**
+   * Queue a run of the render function.
+   * You shouldn't need to call this manually - it should be handled by changes to @property values.
+   * @private
+   */
+  queueRender(): void;
+  /**
+   * Like the standard `element.click()`, but you can influence the behavior with a `sourceEvent`.
+   *
+   * For example, if the `sourceEvent` was a middle click, or has particular keys held down,
+   * components will attempt to produce the desired behavior on links, such as opening the page in the background tab.
+   * @private
+   * @param options
+   */
+  click({sourceEvent}?: ClickOptions): void;
+}
+
+declare class Text extends PreactCustomElement implements TextProps {
+  accessor display: TextProps['display'];
+  accessor fontVariantNumeric: TextProps['fontVariantNumeric'];
+  accessor color: TextProps['color'];
+  accessor tone: TextProps['tone'];
+  accessor type: TextProps['type'];
+  accessor dir: TextProps['dir'];
+  accessor accessibilityVisibility: TextProps['accessibilityVisibility'];
+  constructor();
+}
+declare global {
+  interface HTMLElementTagNameMap {
+    [tagName]: Text;
+  }
+}
+declare module 'preact' {
+  namespace createElement.JSX {
+    interface IntrinsicElements {
+      [tagName]: HTMLAttributes<HTMLElement> & ReactProps;
+    }
+  }
+}
+
+export {Text};
