+++ packages/ui-extensions/src/surfaces/admin/components/shared.d.ts
@@ -1,9 +1,9 @@
-/** VERSION: 0.47.2 **/
+/** VERSION: 0.45.0 **/
+
 /* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable @typescript-eslint/no-namespace */
+
 /* eslint-disable @typescript-eslint/member-ordering */
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference, spaced-comment
-/// <reference lib="DOM" />
+
 type ComponentChildren = any;
 type SizeKeyword =
   | 'small-500'
@@ -1878,24 +1878,6 @@ interface GridProps$1 extends BaseBoxPropsWithRole, GapProps {
     | `${AlignContentKeyword} ${JustifyContentKeyword}`
     | AlignContentKeyword;
 }
-interface GridItemProps$1 extends BaseBoxPropsWithRole {
-  /**
-   * Number of columns the item will span across
-   *
-   * @see https://developer.mozilla.org/en-US/docs/Web/CSS/grid-column
-   *
-   * @default 'auto'
-   */
-  gridColumn?: `span ${number}` | 'auto';
-  /**
-   * Number of rows the item will span across
-   *
-   * @see https://developer.mozilla.org/en-US/docs/Web/CSS/grid-row
-   *
-   * @default 'auto'
-   */
-  gridRow?: `span ${number}` | 'auto';
-}
 interface BaseTypographyProps {
   /**
    * Sets the color of the Typography component, based on the intention of the information being conveyed.
@@ -2515,12 +2497,6 @@ interface StackProps$1 extends BaseBoxPropsWithRole, GapProps {
    */
   alignContent?: AlignContentKeyword;
 }
-interface SwitchProps$1
-  extends GlobalProps,
-    BaseCheckableProps,
-    BasicFieldProps,
-    FieldDetailsProps,
-    FieldErrorProps {}
 interface PaginationProps {
   /**
    * Whether to use pagination controls.
@@ -2882,7 +2858,6 @@ type IconType$1 =
   | 'desktop'
   | 'disabled'
   | 'discount'
-  | 'discount-add'
   | 'discount-code'
   | 'dns-settings'
   | 'dock-floating'
@@ -3034,7 +3009,6 @@ type IconType$1 =
   | 'note-add'
   | 'notification'
   | 'order'
-  | 'order-batches'
   | 'order-draft'
   | 'order-first'
   | 'order-fulfilled'
@@ -3057,7 +3031,6 @@ type IconType$1 =
   | 'page-list'
   | 'page-reference'
   | 'page-remove'
-  | 'page-report'
   | 'page-up'
   | 'pagination-end'
   | 'pagination-start'
@@ -3077,7 +3050,6 @@ type IconType$1 =
   | 'person'
   | 'person-add'
   | 'person-exit'
-  | 'person-list'
   | 'person-lock'
   | 'person-remove'
   | 'person-segment'
@@ -3225,7 +3197,6 @@ type IconType$1 =
   | 'watch'
   | 'wifi'
   | 'work'
-  | 'work-list'
   | 'wrench'
   | 'x'
   | 'x-circle';
@@ -3406,2619 +3377,3 @@ interface AdminPrintActionProps$1 extends GlobalProps {
    */
   src?: string;
 }
-
-export interface IconProps
-  extends Pick<IconProps$1, 'type' | 'tone' | 'color' | 'size'> {
-  type: '' | IconType | 'empty';
-  tone: Extract<
-    IconProps$1['tone'],
-    'auto' | 'neutral' | 'info' | 'success' | 'caution' | 'warning' | 'critical'
-  >;
-  color: Extract<IconProps$1['color'], 'base' | 'subdued'>;
-  size: Extract<IconProps$1['size'], 'small' | 'base'>;
-}
-
-export interface BadgeProps
-  extends Pick<BadgeProps$1, 'color' | 'size' | 'tone'> {
-  color: Extract<BadgeProps$1['color'], 'base' | 'strong'>;
-  icon: IconProps['type'] | '';
-  size: Extract<BadgeProps$1['size'], 'base' | 'large' | 'large-100'>;
-  tone: Extract<
-    BadgeProps$1['tone'],
-    'auto' | 'neutral' | 'info' | 'success' | 'caution' | 'warning' | 'critical'
-  >;
-}
-
-export type Styles = string;
-export type RenderImpl = Omit<ShadowRootInit, 'mode'> & {
-  ShadowRoot: (element: any) => ComponentChild;
-  styles?: Styles;
-};
-export interface ActivationEventEsque {
-  shiftKey: boolean;
-  metaKey: boolean;
-  ctrlKey: boolean;
-  button: number;
-}
-export interface ClickOptions {
-  /**
-   * The event you want to influence the synthetic click.
-   */
-  sourceEvent?: ActivationEventEsque;
-}
-/**
- * Base class for creating custom elements with Preact.
- * While this class could be used in both Node and the browser, the constructor will only be used in the browser.
- * So we give it a type of HTMLElement to avoid typing issues later where it's used, which will only happen in the browser.
- */
-declare const BaseClass: typeof globalThis.HTMLElement;
-declare abstract class PreactCustomElement extends BaseClass {
-  /** @private */
-  static get observedAttributes(): string[];
-  constructor({
-    styles,
-    ShadowRoot: renderFunction,
-    delegatesFocus,
-    ...options
-  }: RenderImpl);
-
-  /** @private */
-  attributeChangedCallback(name: string): void;
-  /** @private */
-  connectedCallback(): void;
-  /** @private */
-  disconnectedCallback(): void;
-  /** @private */
-  adoptedCallback(): void;
-  /**
-   * Queue a run of the render function.
-   * You shouldn't need to call this manually - it should be handled by changes to @property values.
-   * @private
-   */
-  queueRender(): void;
-  /**
-   * Like the standard `element.click()`, but you can influence the behavior with a `sourceEvent`.
-   *
-   * For example, if the `sourceEvent` was a middle click, or has particular keys held down,
-   * components will attempt to produce the desired behavior on links, such as opening the page in the background tab.
-   * @private
-   * @param options
-   */
-  click({sourceEvent}?: ClickOptions): void;
-}
-
-declare class Badge extends PreactCustomElement implements BadgeProps {
-  accessor color: BadgeProps['color'];
-  accessor icon: BadgeProps['icon'];
-  accessor size: BadgeProps['size'];
-  accessor tone: BadgeProps['tone'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$J]: Badge;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$J]: HTMLAttributes<HTMLElement> & BadgeJSXProps;
-    }
-  }
-}
-
-declare const tagName$J = 's-badge';
-export interface BadgeJSXProps
-  extends Partial<BadgeProps>,
-    Pick<BadgeProps$1, 'id'> {}
-
-export type RequiredBannerProps = Required<BannerProps$1>;
-export interface BannerProps
-  extends Pick<RequiredBannerProps, 'heading' | 'dismissible' | 'hidden'> {
-  tone: Extract<
-    RequiredBannerProps['tone'],
-    'critical' | 'warning' | 'success' | 'info'
-  >;
-}
-
-export type CallbackEvent<T extends keyof HTMLElementTagNameMap> = Event & {
-  target: HTMLElementTagNameMap[T];
-};
-export type CallbackEventListener<T extends keyof HTMLElementTagNameMap> =
-  | (EventListener & {
-      (event: CallbackEvent<T>): void;
-    })
-  | null;
-export interface FieldReactProps<T extends keyof HTMLElementTagNameMap> {
-  onInput?: ((event: CallbackEvent<T>) => void) | null;
-  onChange?: ((event: CallbackEvent<T>) => void) | null;
-  onFocus?: ((event: CallbackEvent<T>) => void) | null;
-  onBlur?: ((event: CallbackEvent<T>) => void) | null;
-}
-
-declare class Banner extends PreactCustomElement implements BannerProps {
-  accessor heading: BannerProps['heading'];
-  accessor tone: BannerProps['tone'];
-  accessor hidden: BannerProps['hidden'];
-  accessor dismissible: BannerProps['dismissible'];
-  accessor ondismiss: CallbackEventListener<typeof tagName$I> | null;
-  accessor onafterhide: CallbackEventListener<typeof tagName$I> | null;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$I]: Banner;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$I]: HTMLAttributes<HTMLElement> & BannerJSXProps;
-    }
-  }
-}
-
-declare const tagName$I = 's-banner';
-export interface BannerJSXProps
-  extends Partial<BannerProps>,
-    Pick<BannerProps$1, 'id'> {
-  secondaryActions?: ComponentChild;
-  onDismiss?: ((event: CallbackEvent<typeof tagName$I>) => void) | null;
-  onAfterHide?: ((event: CallbackEvent<typeof tagName$I>) => void) | null;
-}
-
-export type MakeResponsive<T> = T | `@container${string}`;
-
-export type AlignedBox = Required<BoxProps$1>;
-export interface BoxProps
-  extends Pick<
-    AlignedBox,
-    | 'accessibilityRole'
-    | 'background'
-    | 'blockSize'
-    | 'minBlockSize'
-    | 'maxBlockSize'
-    | 'inlineSize'
-    | 'minInlineSize'
-    | 'maxInlineSize'
-    | 'border'
-    | 'borderWidth'
-    | 'borderRadius'
-    | 'borderColor'
-    | 'borderStyle'
-    | 'display'
-    | 'overflow'
-  > {
-  accessibilityRole: AlignedBox['accessibilityRole'];
-  background: Extract<
-    AlignedBox['background'],
-    'transparent' | 'base' | 'subdued' | 'strong'
-  >;
-  blockSize: AlignedBox['blockSize'];
-  minBlockSize: AlignedBox['minBlockSize'];
-  maxBlockSize: AlignedBox['maxBlockSize'];
-  inlineSize: AlignedBox['inlineSize'];
-  minInlineSize: AlignedBox['minInlineSize'];
-  maxInlineSize: AlignedBox['maxInlineSize'];
-  padding: MakeResponsive<AlignedBox['padding']>;
-  paddingBlock: MakeResponsive<AlignedBox['paddingBlock']>;
-  paddingBlockStart: MakeResponsive<AlignedBox['paddingBlockStart']>;
-  paddingBlockEnd: MakeResponsive<AlignedBox['paddingBlockEnd']>;
-  paddingInline: MakeResponsive<AlignedBox['paddingInline']>;
-  paddingInlineStart: MakeResponsive<AlignedBox['paddingInlineStart']>;
-  paddingInlineEnd: MakeResponsive<AlignedBox['paddingInlineEnd']>;
-  border: AlignedBox['border'] | 'none';
-  borderWidth:
-    | MaybeAllValuesShorthandProperty<
-        Extract<
-          AlignedBox['borderWidth'],
-          'small-100' | 'small' | 'base' | 'large' | 'large-100' | 'none'
-        >
-      >
-    | Extract<AlignedBox['borderWidth'], ''>;
-  borderStyle:
-    | MaybeAllValuesShorthandProperty<
-        Extract<AlignedBox['borderStyle'], 'none' | 'solid' | 'dashed'>
-      >
-    | Extract<AlignedBox['borderStyle'], ''>;
-  borderColor: Extract<
-    AlignedBox['borderColor'],
-    'subdued' | 'base' | 'strong' | ''
-  >;
-  borderRadius: MaybeAllValuesShorthandProperty<
-    Extract<
-      AlignedBox['borderRadius'],
-      | 'none'
-      | 'small-200'
-      | 'small-100'
-      | 'small'
-      | 'base'
-      | 'large'
-      | 'large-100'
-      | 'large-200'
-    >
-  >;
-  accessibilityLabel: AlignedBox['accessibilityLabel'];
-  accessibilityVisibility: AlignedBox['accessibilityVisibility'];
-  display: AlignedBox['display'];
-  overflow: AlignedBox['overflow'];
-}
-
-declare class BoxElement extends PreactCustomElement implements BoxProps {
-  constructor(renderImpl: RenderImpl);
-  accessor accessibilityRole: BoxProps['accessibilityRole'];
-  accessor background: BoxProps['background'];
-  accessor blockSize: BoxProps['blockSize'];
-  accessor minBlockSize: BoxProps['minBlockSize'];
-  accessor maxBlockSize: BoxProps['maxBlockSize'];
-  accessor inlineSize: BoxProps['inlineSize'];
-  accessor minInlineSize: BoxProps['minInlineSize'];
-  accessor maxInlineSize: BoxProps['maxInlineSize'];
-  accessor overflow: BoxProps['overflow'];
-  accessor padding: BoxProps['padding'];
-  accessor paddingBlock: BoxProps['paddingBlock'];
-  accessor paddingBlockStart: BoxProps['paddingBlockStart'];
-  accessor paddingBlockEnd: BoxProps['paddingBlockEnd'];
-  accessor paddingInline: BoxProps['paddingInline'];
-  accessor paddingInlineStart: BoxProps['paddingInlineStart'];
-  accessor paddingInlineEnd: BoxProps['paddingInlineEnd'];
-  accessor border: BoxProps['border'];
-  accessor borderWidth: BoxProps['borderWidth'];
-  accessor borderStyle: BoxProps['borderStyle'];
-  accessor borderColor: BoxProps['borderColor'];
-  accessor borderRadius: BoxProps['borderRadius'];
-  accessor accessibilityLabel: BoxProps['accessibilityLabel'];
-  accessor accessibilityVisibility: BoxProps['accessibilityVisibility'];
-  accessor display: BoxProps['display'];
-}
-
-declare class Box extends BoxElement implements BoxProps {
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$H]: Box;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$H]: HTMLAttributes<HTMLElement> & BoxJSXProps;
-    }
-  }
-}
-
-declare const tagName$H = 's-box';
-export interface BoxJSXProps
-  extends Partial<BoxProps>,
-    Pick<BoxProps$1, 'id'> {}
-
-export type ButtonOnlyProps = Extract<
-  ButtonProps$1,
-  {
-    type?: unknown;
-  }
->;
-export type ButtonBaseProps = Required<
-  Pick<
-    ButtonOnlyProps,
-    | 'accessibilityLabel'
-    | 'disabled'
-    | 'command'
-    | 'commandFor'
-    | 'lang'
-    | 'loading'
-    | 'type'
-    | 'tone'
-    | 'variant'
-    | 'target'
-    | 'href'
-    | 'download'
-  >
->;
-export interface ButtonProps extends ButtonBaseProps {
-  tone: Extract<ButtonProps$1['tone'], 'neutral' | 'critical' | 'auto'>;
-  icon: IconProps['type'];
-}
-
-export interface PreactOverlayControlProps
-  extends Required<Pick<InteractionProps, 'commandFor'>> {
-  command: Extract<
-    InteractionProps['command'],
-    '--show' | '--hide' | '--toggle' | '--auto'
-  >;
-}
-
-declare const Button_base: (abstract new (...args: any) => {
-  commandFor: PreactOverlayControlProps['commandFor'];
-  command: PreactOverlayControlProps['command'];
-  '__#49710@#queueRender': (() => void) | undefined;
-  '__#49710@#shadowRoot': ShadowRoot | null;
-  '__#49710@#styles': string;
-  attributeChangedCallback(name: string): void;
-  connectedCallback(): void;
-  disconnectedCallback(): void;
-  adoptedCallback(): void;
-  queueRender(): void;
-  '__#49710@#checkElementPrototype'(): void;
-  click({sourceEvent}?: ClickOptions): void;
-  accessKey: string;
-  readonly accessKeyLabel: string;
-  autocapitalize: string;
-  dir: string;
-  draggable: boolean;
-  hidden: boolean;
-  inert: boolean;
-  innerText: string;
-  lang: string;
-  readonly offsetHeight: number;
-  readonly offsetLeft: number;
-  readonly offsetParent: Element | null;
-  readonly offsetTop: number;
-  readonly offsetWidth: number;
-  outerText: string;
-  popover: string | null;
-  spellcheck: boolean;
-  title: string;
-  translate: boolean;
-  writingSuggestions: string;
-  attachInternals(): ElementInternals;
-  hidePopover(): void;
-  showPopover(): void;
-  togglePopover(force?: boolean): boolean;
-  addEventListener<K extends keyof HTMLElementEventMap>(
-    type: K,
-    listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any,
-    options?: boolean | AddEventListenerOptions,
-  ): void;
-  addEventListener(
-    type: string,
-    listener: EventListenerOrEventListenerObject,
-    options?: boolean | AddEventListenerOptions,
-  ): void;
-  removeEventListener<K extends keyof HTMLElementEventMap>(
-    type: K,
-    listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any,
-    options?: boolean | EventListenerOptions,
-  ): void;
-  removeEventListener(
-    type: string,
-    listener: EventListenerOrEventListenerObject,
-    options?: boolean | EventListenerOptions,
-  ): void;
-  readonly attributes: NamedNodeMap;
-  readonly classList: DOMTokenList;
-  className: string;
-  readonly clientHeight: number;
-  readonly clientLeft: number;
-  readonly clientTop: number;
-  readonly clientWidth: number;
-  readonly currentCSSZoom: number;
-  id: string;
-  innerHTML: string;
-  readonly localName: string;
-  readonly namespaceURI: string | null;
-  onfullscreenchange: ((this: Element, ev: Event) => any) | null;
-  onfullscreenerror: ((this: Element, ev: Event) => any) | null;
-  outerHTML: string;
-  readonly ownerDocument: Document;
-  readonly part: DOMTokenList;
-  readonly prefix: string | null;
-  readonly scrollHeight: number;
-  scrollLeft: number;
-  scrollTop: number;
-  readonly scrollWidth: number;
-  readonly shadowRoot: ShadowRoot | null;
-  slot: string;
-  readonly tagName: string;
-  attachShadow(init: ShadowRootInit): ShadowRoot;
-  checkVisibility(options?: CheckVisibilityOptions): boolean;
-  closest<K extends keyof HTMLElementTagNameMap>(
-    selector: K,
-  ): HTMLElementTagNameMap[K] | null;
-  closest<K extends keyof SVGElementTagNameMap>(
-    selector: K,
-  ): SVGElementTagNameMap[K] | null;
-  closest<K extends keyof MathMLElementTagNameMap>(
-    selector: K,
-  ): MathMLElementTagNameMap[K] | null;
-  closest<E extends Element = Element>(selectors: string): E | null;
-  computedStyleMap(): StylePropertyMapReadOnly;
-  getAttribute(qualifiedName: string): string | null;
-  getAttributeNS(namespace: string | null, localName: string): string | null;
-  getAttributeNames(): string[];
-  getAttributeNode(qualifiedName: string): Attr | null;
-  getAttributeNodeNS(namespace: string | null, localName: string): Attr | null;
-  getBoundingClientRect(): DOMRect;
-  getClientRects(): DOMRectList;
-  getElementsByClassName(classNames: string): HTMLCollectionOf<Element>;
-  getElementsByTagName<K extends keyof HTMLElementTagNameMap>(
-    qualifiedName: K,
-  ): HTMLCollectionOf<HTMLElementTagNameMap[K]>;
-  getElementsByTagName<K extends keyof SVGElementTagNameMap>(
-    qualifiedName: K,
-  ): HTMLCollectionOf<SVGElementTagNameMap[K]>;
-  getElementsByTagName<K extends keyof MathMLElementTagNameMap>(
-    qualifiedName: K,
-  ): HTMLCollectionOf<MathMLElementTagNameMap[K]>;
-  getElementsByTagName<K extends keyof HTMLElementDeprecatedTagNameMap>(
-    qualifiedName: K,
-  ): HTMLCollectionOf<HTMLElementDeprecatedTagNameMap[K]>;
-  getElementsByTagName(qualifiedName: string): HTMLCollectionOf<Element>;
-  getElementsByTagNameNS(
-    namespaceURI: 'http://www.w3.org/1999/xhtml',
-    localName: string,
-  ): HTMLCollectionOf<HTMLElement>;
-  getElementsByTagNameNS(
-    namespaceURI: 'http://www.w3.org/2000/svg',
-    localName: string,
-  ): HTMLCollectionOf<SVGElement>;
-  getElementsByTagNameNS(
-    namespaceURI: 'http://www.w3.org/1998/Math/MathML',
-    localName: string,
-  ): HTMLCollectionOf<MathMLElement>;
-  getElementsByTagNameNS(
-    namespace: string | null,
-    localName: string,
-  ): HTMLCollectionOf<Element>;
-  getHTML(options?: GetHTMLOptions): string;
-  hasAttribute(qualifiedName: string): boolean;
-  hasAttributeNS(namespace: string | null, localName: string): boolean;
-  hasAttributes(): boolean;
-  hasPointerCapture(pointerId: number): boolean;
-  insertAdjacentElement(
-    where: InsertPosition,
-    element: Element,
-  ): Element | null;
-  insertAdjacentHTML(position: InsertPosition, string: string): void;
-  insertAdjacentText(where: InsertPosition, data: string): void;
-  matches(selectors: string): boolean;
-  releasePointerCapture(pointerId: number): void;
-  removeAttribute(qualifiedName: string): void;
-  removeAttributeNS(namespace: string | null, localName: string): void;
-  removeAttributeNode(attr: Attr): Attr;
-  requestFullscreen(options?: FullscreenOptions): Promise<void>;
-  requestPointerLock(options?: PointerLockOptions): Promise<void>;
-  scroll(options?: ScrollToOptions): void;
-  scroll(x: number, y: number): void;
-  scrollBy(options?: ScrollToOptions): void;
-  scrollBy(x: number, y: number): void;
-  scrollIntoView(arg?: boolean | ScrollIntoViewOptions): void;
-  scrollTo(options?: ScrollToOptions): void;
-  scrollTo(x: number, y: number): void;
-  setAttribute(qualifiedName: string, value: string): void;
-  setAttributeNS(
-    namespace: string | null,
-    qualifiedName: string,
-    value: string,
-  ): void;
-  setAttributeNode(attr: Attr): Attr | null;
-  setAttributeNodeNS(attr: Attr): Attr | null;
-  setHTMLUnsafe(html: string): void;
-  setPointerCapture(pointerId: number): void;
-  toggleAttribute(qualifiedName: string, force?: boolean): boolean;
-  webkitMatchesSelector(selectors: string): boolean;
-  readonly baseURI: string;
-  readonly childNodes: NodeListOf<ChildNode>;
-  readonly firstChild: ChildNode | null;
-  readonly isConnected: boolean;
-  readonly lastChild: ChildNode | null;
-  readonly nextSibling: ChildNode | null;
-  readonly nodeName: string;
-  readonly nodeType: number;
-  nodeValue: string | null;
-  readonly parentElement: HTMLElement | null;
-  readonly parentNode: ParentNode | null;
-  readonly previousSibling: ChildNode | null;
-  textContent: string | null;
-  appendChild<T extends Node>(node: T): T;
-  cloneNode(deep?: boolean): Node;
-  compareDocumentPosition(other: Node): number;
-  contains(other: Node | null): boolean;
-  getRootNode(options?: GetRootNodeOptions): Node;
-  hasChildNodes(): boolean;
-  insertBefore<T extends Node>(node: T, child: Node | null): T;
-  isDefaultNamespace(namespace: string | null): boolean;
-  isEqualNode(otherNode: Node | null): boolean;
-  isSameNode(otherNode: Node | null): boolean;
-  lookupNamespaceURI(prefix: string | null): string | null;
-  lookupPrefix(namespace: string | null): string | null;
-  normalize(): void;
-  removeChild<T extends Node>(child: T): T;
-  replaceChild<T extends Node>(node: Node, child: T): T;
-  readonly ELEMENT_NODE: 1;
-  readonly ATTRIBUTE_NODE: 2;
-  readonly TEXT_NODE: 3;
-  readonly CDATA_SECTION_NODE: 4;
-  readonly ENTITY_REFERENCE_NODE: 5;
-  readonly ENTITY_NODE: 6;
-  readonly PROCESSING_INSTRUCTION_NODE: 7;
-  readonly COMMENT_NODE: 8;
-  readonly DOCUMENT_NODE: 9;
-  readonly DOCUMENT_TYPE_NODE: 10;
-  readonly DOCUMENT_FRAGMENT_NODE: 11;
-  readonly NOTATION_NODE: 12;
-  readonly DOCUMENT_POSITION_DISCONNECTED: 1;
-  readonly DOCUMENT_POSITION_PRECEDING: 2;
-  readonly DOCUMENT_POSITION_FOLLOWING: 4;
-  readonly DOCUMENT_POSITION_CONTAINS: 8;
-  readonly DOCUMENT_POSITION_CONTAINED_BY: 16;
-  readonly DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC: 32;
-  dispatchEvent(event: Event): boolean;
-  dispatchEvent(event: Event): boolean;
-  ariaAtomic: string | null;
-  ariaAutoComplete: string | null;
-  ariaBrailleLabel: string | null;
-  ariaBrailleRoleDescription: string | null;
-  ariaBusy: string | null;
-  ariaChecked: string | null;
-  ariaColCount: string | null;
-  ariaColIndex: string | null;
-  ariaColIndexText: string | null;
-  ariaColSpan: string | null;
-  ariaCurrent: string | null;
-  ariaDescription: string | null;
-  ariaDisabled: string | null;
-  ariaExpanded: string | null;
-  ariaHasPopup: string | null;
-  ariaHidden: string | null;
-  ariaInvalid: string | null;
-  ariaKeyShortcuts: string | null;
-  ariaLabel: string | null;
-  ariaLevel: string | null;
-  ariaLive: string | null;
-  ariaModal: string | null;
-  ariaMultiLine: string | null;
-  ariaMultiSelectable: string | null;
-  ariaOrientation: string | null;
-  ariaPlaceholder: string | null;
-  ariaPosInSet: string | null;
-  ariaPressed: string | null;
-  ariaReadOnly: string | null;
-  ariaRequired: string | null;
-  ariaRoleDescription: string | null;
-  ariaRowCount: string | null;
-  ariaRowIndex: string | null;
-  ariaRowIndexText: string | null;
-  ariaRowSpan: string | null;
-  ariaSelected: string | null;
-  ariaSetSize: string | null;
-  ariaSort: string | null;
-  ariaValueMax: string | null;
-  ariaValueMin: string | null;
-  ariaValueNow: string | null;
-  ariaValueText: string | null;
-  role: string | null;
-  animate(
-    keyframes: Keyframe[] | PropertyIndexedKeyframes | null,
-    options?: number | KeyframeAnimationOptions,
-  ): Animation;
-  getAnimations(options?: GetAnimationsOptions): Animation[];
-  after(...nodes: (Node | string)[]): void;
-  before(...nodes: (Node | string)[]): void;
-  remove(): void;
-  replaceWith(...nodes: (Node | string)[]): void;
-  readonly nextElementSibling: Element | null;
-  readonly previousElementSibling: Element | null;
-  readonly childElementCount: number;
-  readonly children: HTMLCollection;
-  readonly firstElementChild: Element | null;
-  readonly lastElementChild: Element | null;
-  append(...nodes: (Node | string)[]): void;
-  prepend(...nodes: (Node | string)[]): void;
-  querySelector<K extends keyof HTMLElementTagNameMap>(
-    selectors: K,
-  ): HTMLElementTagNameMap[K] | null;
-  querySelector<K extends keyof SVGElementTagNameMap>(
-    selectors: K,
-  ): SVGElementTagNameMap[K] | null;
-  querySelector<K extends keyof MathMLElementTagNameMap>(
-    selectors: K,
-  ): MathMLElementTagNameMap[K] | null;
-  querySelector<K extends keyof HTMLElementDeprecatedTagNameMap>(
-    selectors: K,
-  ): HTMLElementDeprecatedTagNameMap[K] | null;
-  querySelector<E extends Element = Element>(selectors: string): E | null;
-  querySelectorAll<K extends keyof HTMLElementTagNameMap>(
-    selectors: K,
-  ): NodeListOf<HTMLElementTagNameMap[K]>;
-  querySelectorAll<K extends keyof SVGElementTagNameMap>(
-    selectors: K,
-  ): NodeListOf<SVGElementTagNameMap[K]>;
-  querySelectorAll<K extends keyof MathMLElementTagNameMap>(
-    selectors: K,
-  ): NodeListOf<MathMLElementTagNameMap[K]>;
-  querySelectorAll<K extends keyof HTMLElementDeprecatedTagNameMap>(
-    selectors: K,
-  ): NodeListOf<HTMLElementDeprecatedTagNameMap[K]>;
-  querySelectorAll<E extends Element = Element>(
-    selectors: string,
-  ): NodeListOf<E>;
-  replaceChildren(...nodes: (Node | string)[]): void;
-  readonly assignedSlot: HTMLSlotElement | null;
-  readonly attributeStyleMap: StylePropertyMap;
-  readonly style: CSSStyleDeclaration;
-  contentEditable: string;
-  enterKeyHint: string;
-  inputMode: string;
-  readonly isContentEditable: boolean;
-  onabort: ((this: GlobalEventHandlers, ev: UIEvent) => any) | null;
-  onanimationcancel:
-    | ((this: GlobalEventHandlers, ev: AnimationEvent) => any)
-    | null;
-  onanimationend:
-    | ((this: GlobalEventHandlers, ev: AnimationEvent) => any)
-    | null;
-  onanimationiteration:
-    | ((this: GlobalEventHandlers, ev: AnimationEvent) => any)
-    | null;
-  onanimationstart:
-    | ((this: GlobalEventHandlers, ev: AnimationEvent) => any)
-    | null;
-  onauxclick: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onbeforeinput: ((this: GlobalEventHandlers, ev: InputEvent) => any) | null;
-  onbeforetoggle: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onblur: ((this: GlobalEventHandlers, ev: FocusEvent) => any) | null;
-  oncancel: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncanplay: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncanplaythrough: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onchange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onclick: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onclose: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncontextlost: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncontextmenu: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  oncontextrestored: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncopy: ((this: GlobalEventHandlers, ev: ClipboardEvent) => any) | null;
-  oncuechange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oncut: ((this: GlobalEventHandlers, ev: ClipboardEvent) => any) | null;
-  ondblclick: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  ondrag: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondragend: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondragenter: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondragleave: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondragover: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondragstart: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondrop: ((this: GlobalEventHandlers, ev: DragEvent) => any) | null;
-  ondurationchange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onemptied: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onended: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onerror: OnErrorEventHandler;
-  onfocus: ((this: GlobalEventHandlers, ev: FocusEvent) => any) | null;
-  onformdata: ((this: GlobalEventHandlers, ev: FormDataEvent) => any) | null;
-  ongotpointercapture:
-    | ((this: GlobalEventHandlers, ev: PointerEvent) => any)
-    | null;
-  oninput: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  oninvalid: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onkeydown: ((this: GlobalEventHandlers, ev: KeyboardEvent) => any) | null;
-  onkeypress: ((this: GlobalEventHandlers, ev: KeyboardEvent) => any) | null;
-  onkeyup: ((this: GlobalEventHandlers, ev: KeyboardEvent) => any) | null;
-  onload: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onloadeddata: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onloadedmetadata: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onloadstart: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onlostpointercapture:
-    | ((this: GlobalEventHandlers, ev: PointerEvent) => any)
-    | null;
-  onmousedown: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmouseenter: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmouseleave: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmousemove: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmouseout: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmouseover: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onmouseup: ((this: GlobalEventHandlers, ev: MouseEvent) => any) | null;
-  onpaste: ((this: GlobalEventHandlers, ev: ClipboardEvent) => any) | null;
-  onpause: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onplay: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onplaying: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onpointercancel:
-    | ((this: GlobalEventHandlers, ev: PointerEvent) => any)
-    | null;
-  onpointerdown: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointerenter: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointerleave: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointermove: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointerout: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointerover: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onpointerup: ((this: GlobalEventHandlers, ev: PointerEvent) => any) | null;
-  onprogress: ((this: GlobalEventHandlers, ev: ProgressEvent) => any) | null;
-  onratechange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onreset: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onresize: ((this: GlobalEventHandlers, ev: UIEvent) => any) | null;
-  onscroll: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onscrollend: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onsecuritypolicyviolation:
-    | ((this: GlobalEventHandlers, ev: SecurityPolicyViolationEvent) => any)
-    | null;
-  onseeked: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onseeking: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onselect: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onselectionchange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onselectstart: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onslotchange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onstalled: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onsubmit: ((this: GlobalEventHandlers, ev: SubmitEvent) => any) | null;
-  onsuspend: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  ontimeupdate: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  ontoggle: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  ontouchcancel?:
-    | ((this: GlobalEventHandlers, ev: TouchEvent) => any)
-    | null
-    | undefined;
-  ontouchend?:
-    | ((this: GlobalEventHandlers, ev: TouchEvent) => any)
-    | null
-    | undefined;
-  ontouchmove?:
-    | ((this: GlobalEventHandlers, ev: TouchEvent) => any)
-    | null
-    | undefined;
-  ontouchstart?:
-    | ((this: GlobalEventHandlers, ev: TouchEvent) => any)
-    | null
-    | undefined;
-  ontransitioncancel:
-    | ((this: GlobalEventHandlers, ev: TransitionEvent) => any)
-    | null;
-  ontransitionend:
-    | ((this: GlobalEventHandlers, ev: TransitionEvent) => any)
-    | null;
-  ontransitionrun:
-    | ((this: GlobalEventHandlers, ev: TransitionEvent) => any)
-    | null;
-  ontransitionstart:
-    | ((this: GlobalEventHandlers, ev: TransitionEvent) => any)
-    | null;
-  onvolumechange: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onwaiting: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onwebkitanimationend: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onwebkitanimationiteration:
-    | ((this: GlobalEventHandlers, ev: Event) => any)
-    | null;
-  onwebkitanimationstart:
-    | ((this: GlobalEventHandlers, ev: Event) => any)
-    | null;
-  onwebkittransitionend: ((this: GlobalEventHandlers, ev: Event) => any) | null;
-  onwheel: ((this: GlobalEventHandlers, ev: WheelEvent) => any) | null;
-  autofocus: boolean;
-  readonly dataset: DOMStringMap;
-  nonce?: string;
-  tabIndex: number;
-  blur(): void;
-  focus(options?: FocusOptions): void;
-}) &
-  typeof PreactCustomElement;
-declare class Button extends Button_base implements ButtonProps {
-  accessor disabled: ButtonProps['disabled'];
-  accessor icon: ButtonProps['icon'];
-  accessor loading: ButtonProps['loading'];
-  accessor variant: ButtonProps['variant'];
-  accessor tone: ButtonProps['tone'];
-  accessor target: ButtonProps['target'];
-  accessor href: ButtonProps['href'];
-  accessor download: ButtonProps['download'];
-  accessor onclick: CallbackEventListener<typeof tagName$G> | null;
-  accessor onblur: CallbackEventListener<typeof tagName$G> | null;
-  accessor onfocus: CallbackEventListener<typeof tagName$G> | null;
-  accessor type: ButtonProps['type'];
-  accessor accessibilityLabel: ButtonProps['accessibilityLabel'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$G]: Button;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$G]: HTMLAttributes<HTMLElement> & ButtonJSXProps;
-    }
-  }
-}
-
-declare const tagName$G = 's-button';
-export interface ButtonJSXProps
-  extends Partial<ButtonProps>,
-    Pick<ButtonProps$1, 'id'> {
-  onClick?: ((event: CallbackEvent<typeof tagName$G>) => void) | null;
-  onFocus?: ((event: CallbackEvent<typeof tagName$G>) => void) | null;
-  onBlur?: ((event: CallbackEvent<typeof tagName$G>) => void) | null;
-}
-
-declare const internals: unique symbol;
-export type PreactInputProps = Required<
-  Pick<TextFieldProps$1, 'disabled' | 'id' | 'name' | 'value'>
->;
-declare class PreactInputElement
-  extends PreactCustomElement
-  implements PreactInputProps
-{
-  static formAssociated: boolean;
-  /** @private */
-  [internals]: ElementInternals;
-  protected getDefaultValue(): string;
-  accessor onchange: CallbackEventListener<'input'>;
-  accessor oninput: CallbackEventListener<'input'>;
-  accessor disabled: PreactInputProps['disabled'];
-  accessor id: PreactInputProps['id'];
-  accessor name: PreactInputProps['name'];
-  get value(): PreactInputProps['value'];
-  set value(value: PreactInputProps['value']);
-  constructor(renderImpl: RenderImpl);
-}
-
-export type CheckboxProps = PreactInputProps &
-  Required<
-    Pick<
-      CheckboxProps$1,
-      | 'accessibilityLabel'
-      | 'checked'
-      | 'details'
-      | 'error'
-      | 'indeterminate'
-      | 'label'
-      | 'required'
-      | 'defaultChecked'
-      | 'defaultIndeterminate'
-    >
-  >;
-
-export type PreactCheckboxProps = Required<
-  Pick<
-    CheckboxProps,
-    | 'accessibilityLabel'
-    | 'checked'
-    | 'defaultChecked'
-    | 'details'
-    | 'error'
-    | 'label'
-    | 'required'
-    | 'name'
-    | 'disabled'
-    | 'value'
-  >
->;
-declare class PreactCheckboxElement
-  extends PreactInputElement
-  implements PreactCheckboxProps
-{
-  get checked(): boolean;
-  set checked(checked: CheckboxProps['checked']);
-  get value(): string;
-  set value(value: string);
-  accessor defaultChecked: CheckboxProps['defaultChecked'];
-  accessor accessibilityLabel: CheckboxProps['accessibilityLabel'];
-  accessor details: CheckboxProps['details'];
-  accessor error: CheckboxProps['error'];
-  get indeterminate(): CheckboxProps['indeterminate'];
-  set indeterminate(indeterminate: CheckboxProps['indeterminate']);
-  accessor defaultIndeterminate: CheckboxProps['defaultIndeterminate'];
-  accessor label: CheckboxProps['label'];
-  accessor required: CheckboxProps['required'];
-  formResetCallback(): void;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$F]: Checkbox;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$F]: HTMLAttributes<HTMLElement> & CheckboxJSXProps;
-    }
-  }
-}
-
-declare const tagName$F = 's-checkbox';
-export interface CheckboxJSXProps
-  extends Partial<CheckboxProps>,
-    Pick<CheckboxProps$1, 'id'> {
-  onChange?: ((event: CallbackEvent<typeof tagName$F>) => void) | null;
-  onInput?: ((event: CallbackEvent<typeof tagName$F>) => void) | null;
-}
-
-export interface ChoiceProps
-  extends Required<
-    Pick<
-      ChoiceProps$1,
-      | 'selected'
-      | 'defaultSelected'
-      | 'disabled'
-      | 'accessibilityLabel'
-      | 'value'
-      | 'label'
-      | 'details'
-    >
-  > {}
-
-declare class Choice extends PreactCustomElement implements ChoiceProps {
-  accessor disabled: ChoiceProps['disabled'];
-  accessor details: ChoiceProps['details'];
-  get selected(): boolean;
-  set selected(selected: ChoiceProps['selected']);
-  accessor value: ChoiceProps['value'];
-  accessor accessibilityLabel: ChoiceProps['accessibilityLabel'];
-  accessor label: ChoiceProps['label'];
-  accessor defaultSelected: ChoiceProps['defaultSelected'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$E]: Choice;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$E]: HTMLAttributes<HTMLElement> & ChoiceJSXProps;
-    }
-  }
-}
-
-declare const tagName$E = 's-choice';
-export interface ChoiceJSXProps
-  extends Partial<ChoiceProps>,
-    Pick<ChoiceProps$1, 'id'> {}
-
-export interface ChoiceListProps
-  extends Required<
-    Pick<
-      ChoiceListProps$1,
-      'multiple' | 'disabled' | 'error' | 'details' | 'values' | 'name'
-    >
-  > {}
-
-declare class ChoiceList extends PreactInputElement implements ChoiceListProps {
-  accessor error: ChoiceListProps['error'];
-  accessor details: ChoiceListProps['details'];
-  accessor multiple: ChoiceListProps['multiple'];
-  get values(): ChoiceListProps['values'];
-  set values(values: ChoiceListProps['values']);
-  formResetCallback(): void;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$D]: ChoiceList;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$D]: HTMLAttributes<HTMLElement> & ChoiceListJSXProps;
-    }
-  }
-}
-
-declare const tagName$D = 's-choice-list';
-export interface ChoiceListJSXProps
-  extends Partial<ChoiceListProps>,
-    Pick<ChoiceListProps$1, 'id'> {
-  onChange?: ((event: CallbackEvent<typeof tagName$D>) => void) | null;
-  onInput?: ((event: CallbackEvent<typeof tagName$D>) => void) | null;
-}
-
-export type ClickableBaseProps = Required<
-  Pick<
-    ClickableProps$1,
-    | 'command'
-    | 'commandFor'
-    | 'disabled'
-    | 'download'
-    | 'href'
-    | 'lang'
-    | 'loading'
-    | 'overflow'
-    | 'target'
-    | 'type'
-  >
->;
-export interface ClickableProps
-  extends Required<BoxProps>,
-    ClickableBaseProps {}
-
-declare const Clickable_base: (abstract new (
-  renderImpl: RenderImpl,
-) => BoxElement & PreactOverlayControlProps) &
-  Pick<typeof BoxElement, 'observedAttributes' | 'prototype'>;
-declare class Clickable extends Clickable_base implements ClickableProps {
-  accessor disabled: ClickableProps['disabled'];
-  accessor loading: ClickableProps['loading'];
-  accessor target: ClickableProps['target'];
-  accessor href: ClickableProps['href'];
-  accessor download: ClickableProps['download'];
-  accessor onclick: CallbackEventListener<typeof tagName$C> | null;
-  accessor onblur: CallbackEventListener<typeof tagName$C> | null;
-  accessor onfocus: CallbackEventListener<typeof tagName$C> | null;
-  accessor type: ClickableProps['type'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$C]: Clickable;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$C]: HTMLAttributes<HTMLElement> & ClickableJSXProps;
-    }
-  }
-}
-
-declare const tagName$C = 's-clickable';
-export interface ClickableJSXProps
-  extends Partial<ClickableProps>,
-    Pick<ClickableProps$1, 'id'> {
-  onClick?: ((event: CallbackEvent<typeof tagName$C>) => void) | null;
-  onFocus?: ((event: CallbackEvent<typeof tagName$C>) => void) | null;
-  onBlur?: ((event: CallbackEvent<typeof tagName$C>) => void) | null;
-}
-
-export interface DividerProps
-  extends Pick<DividerProps$1, 'direction' | 'color'> {
-  direction: Extract<DividerProps$1['direction'], 'inline' | 'block'>;
-  color: Extract<DividerProps$1['color'], 'base' | 'strong'>;
-}
-
-declare class Divider extends PreactCustomElement implements DividerProps {
-  accessor direction: DividerProps['direction'];
-  accessor color: DividerProps['color'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$B]: Divider;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$B]: HTMLAttributes<HTMLElement> & DividerJSXProps;
-    }
-  }
-}
-
-declare const tagName$B = 's-divider';
-export interface DividerJSXProps
-  extends Partial<DividerProps>,
-    Pick<DividerProps$1, 'id'> {}
-
-export type PreactFieldProps<Autocomplete extends string = string> =
-  PreactInputProps &
-    Required<
-      Pick<
-        TextFieldProps$1,
-        | 'defaultValue'
-        | 'details'
-        | 'error'
-        | 'label'
-        | 'labelAccessibilityVisibility'
-        | 'placeholder'
-        | 'readOnly'
-        | 'required'
-      >
-    > & {
-      /**
-       * A hint as to the intended content of the field.
-       *
-       * When set to `on` (the default), this property indicates that the field should support
-       * autofill, but you do not have any more semantic information on the intended
-       * contents.
-       *
-       * When set to `off`, you are indicating that this field contains sensitive
-       * information, or contents that are never saved, like one-time codes.
-       *
-       * Alternatively, you can provide value which describes the
-       * specific data you would like to be entered into this field during autofill.
-       *
-       * @see Learn more about the set of {@link https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#autofill-detail-tokens|autocomplete values} supported in browsers.
-       *
-       * @default 'tel' for PhoneField
-       * @default 'email' for EmailField
-       * @default 'url' for URLField
-       * @default 'on' for everything else
-       */
-      autocomplete: Autocomplete;
-    };
-declare class PreactFieldElement<Autocomplete extends string = string>
-  extends PreactInputElement
-  implements PreactFieldProps<Autocomplete>
-{
-  accessor onblur: CallbackEventListener<'input'>;
-  accessor onfocus: CallbackEventListener<'input'>;
-  accessor autocomplete: PreactFieldProps<Autocomplete>['autocomplete'];
-  accessor defaultValue: PreactFieldProps['defaultValue'];
-  accessor details: PreactFieldProps['details'];
-  accessor error: PreactFieldProps['error'];
-  accessor label: PreactFieldProps['label'];
-  accessor labelAccessibilityVisibility: PreactFieldProps['labelAccessibilityVisibility'];
-  accessor placeholder: PreactFieldProps['placeholder'];
-  accessor readOnly: PreactFieldProps['readOnly'];
-  accessor required: PreactFieldProps['required'];
-  protected getDefaultValue(): string;
-  /**
-   * Global keyboard event handlers for things like key bindings typically
-   * ignore keystrokes originating from within input elements. Unfortunately,
-   * these never account for a Custom Element being the input element.
-   *
-   * To fix this, we spoof getAttribute & hasAttribute to make a PreactFieldElement
-   * appear as a contentEditable "input" when it contains a focused input element.
-   */
-  getAttribute(qualifiedName: string): string | null;
-  hasAttribute(qualifiedName: string): boolean;
-  /**
-   * Checks if the shadow tree contains a focused input (input, textarea, select, <x contentEditable>).
-   * Note: this does _not_ return true for focussed non-field form elements like buttons.
-   */
-  get isContentEditable(): boolean;
-  formResetCallback(): void;
-  connectedCallback(): void;
-  constructor(renderImpl: RenderImpl);
-}
-
-export type EmailFieldProps = PreactFieldProps<
-  Required<EmailFieldProps$1>['autocomplete']
-> &
-  Required<Pick<EmailFieldProps$1, 'maxLength' | 'minLength'>>;
-
-declare class EmailField
-  extends PreactFieldElement<EmailFieldProps['autocomplete']>
-  implements EmailFieldProps
-{
-  accessor autocomplete: EmailFieldProps['autocomplete'];
-  accessor maxLength: EmailFieldProps['maxLength'];
-  accessor minLength: EmailFieldProps['minLength'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$A]: EmailField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$A]: HTMLAttributes<HTMLElement> & EmailFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$A = 's-email-field';
-export interface EmailFieldJSXProps
-  extends Partial<Omit<EmailFieldProps, 'accessory'>>,
-    Pick<EmailFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$A> {}
-
-export type RequiredAlignedProps = Required<GridProps$1>;
-export interface GridProps
-  extends BoxProps,
-    Required<
-      Pick<
-        GridProps$1,
-        | 'gridTemplateColumns'
-        | 'gridTemplateRows'
-        | 'alignItems'
-        | 'justifyItems'
-        | 'placeItems'
-        | 'alignContent'
-        | 'justifyContent'
-        | 'placeContent'
-      >
-    > {
-  gridTemplateColumns: RequiredAlignedProps['gridTemplateColumns'];
-  gridTemplateRows: RequiredAlignedProps['gridTemplateRows'];
-  alignItems: RequiredAlignedProps['alignItems'];
-  justifyContent: RequiredAlignedProps['justifyContent'];
-  placeContent: RequiredAlignedProps['placeContent'];
-  rowGap: MakeResponsive<RequiredAlignedProps['rowGap']>;
-  columnGap: MakeResponsive<RequiredAlignedProps['columnGap']>;
-  gap: MakeResponsive<RequiredAlignedProps['gap']>;
-}
-
-declare class Grid extends BoxElement implements GridProps {
-  constructor();
-  accessor gridTemplateColumns: GridProps['gridTemplateColumns'];
-  accessor gridTemplateRows: GridProps['gridTemplateRows'];
-  accessor justifyItems: GridProps['justifyItems'];
-  accessor alignItems: GridProps['alignItems'];
-  accessor placeItems: GridProps['placeItems'];
-  accessor justifyContent: GridProps['justifyContent'];
-  accessor alignContent: GridProps['alignContent'];
-  accessor placeContent: GridProps['placeContent'];
-  accessor gap: GridProps['gap'];
-  accessor rowGap: GridProps['rowGap'];
-  accessor columnGap: GridProps['columnGap'];
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$z]: Grid;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$z]: HTMLAttributes<HTMLElement> & GridJSXProps;
-    }
-  }
-}
-
-declare const tagName$z = 's-grid';
-export interface GridJSXProps
-  extends Partial<GridProps>,
-    Pick<GridProps$1, 'id'> {}
-
-export type RequiredGridItemProps = Required<GridItemProps$1>;
-export interface GridItemProps
-  extends BoxProps,
-    Required<Pick<GridItemProps$1, 'gridColumn' | 'gridRow'>> {
-  gridColumn: RequiredGridItemProps['gridColumn'];
-  gridRow: RequiredGridItemProps['gridRow'];
-}
-
-declare class GridItem extends BoxElement implements GridItemProps {
-  accessor gridColumn: GridItemProps['gridColumn'];
-  accessor gridRow: GridItemProps['gridRow'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$y]: GridItem;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$y]: HTMLAttributes<HTMLElement> & GridItemJSXProps;
-    }
-  }
-}
-
-declare const tagName$y = 's-grid-item';
-export interface GridItemJSXProps
-  extends Partial<GridItemProps>,
-    Pick<GridItemProps$1, 'id'> {}
-
-export type RequiredHeadingProps = Required<HeadingProps$1>;
-export interface HeadingProps
-  extends Pick<
-    HeadingProps$1,
-    'accessibilityRole' | 'accessibilityVisibility' | 'lineClamp'
-  > {
-  accessibilityRole: RequiredHeadingProps['accessibilityRole'];
-  accessibilityVisibility: RequiredHeadingProps['accessibilityVisibility'];
-  lineClamp: RequiredHeadingProps['lineClamp'];
-}
-
-declare class Heading extends PreactCustomElement implements HeadingProps {
-  accessor accessibilityRole: HeadingProps['accessibilityRole'];
-  accessor lineClamp: HeadingProps['lineClamp'];
-  accessor accessibilityVisibility: HeadingProps['accessibilityVisibility'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$x]: Heading;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$x]: HTMLAttributes<HTMLElement> & HeadingJSXProps;
-    }
-  }
-}
-
-declare const tagName$x = 's-heading';
-export interface HeadingJSXProps
-  extends Partial<HeadingProps>,
-    Pick<HeadingProps$1, 'id'> {}
-
-declare class Icon extends PreactCustomElement implements IconProps {
-  accessor color: IconProps['color'];
-  accessor tone: IconProps['tone'];
-  accessor type: IconProps['type'];
-  accessor size: IconProps['size'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$w]: Icon;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$w]: Omit<HTMLAttributes<HTMLElement>, 'size'> & IconJSXProps;
-    }
-  }
-}
-
-declare const tagName$w = 's-icon';
-export interface IconJSXProps
-  extends Partial<IconProps>,
-    Pick<IconProps$1, 'id'> {}
-
-export interface ImageProps
-  extends Required<
-      Pick<
-        ImageProps$1,
-        | 'alt'
-        | 'loading'
-        | 'src'
-        | 'accessibilityRole'
-        | 'inlineSize'
-        | 'srcSet'
-        | 'sizes'
-        | 'aspectRatio'
-        | 'objectFit'
-      >
-    >,
-    Required<
-      Pick<
-        BoxProps,
-        | 'border'
-        | 'borderColor'
-        | 'borderRadius'
-        | 'borderStyle'
-        | 'borderWidth'
-      >
-    > {
-  border: BoxProps['border'];
-  borderColor: BoxProps['borderColor'];
-  borderRadius: BoxProps['borderRadius'];
-  borderStyle: BoxProps['borderStyle'];
-  borderWidth: BoxProps['borderWidth'];
-}
-
-declare class Image extends PreactCustomElement implements ImageProps {
-  accessor src: ImageProps['src'];
-  accessor srcSet: ImageProps['srcSet'];
-  accessor sizes: ImageProps['sizes'];
-  accessor alt: ImageProps['alt'];
-  accessor aspectRatio: ImageProps['aspectRatio'];
-  accessor objectFit: ImageProps['objectFit'];
-  accessor loading: ImageProps['loading'];
-  accessor accessibilityRole: ImageProps['accessibilityRole'];
-  accessor inlineSize: ImageProps['inlineSize'];
-  accessor border: ImageProps['border'];
-  accessor borderWidth: ImageProps['borderWidth'];
-  accessor borderStyle: ImageProps['borderStyle'];
-  accessor borderColor: ImageProps['borderColor'];
-  accessor borderRadius: ImageProps['borderRadius'];
-  accessor onload: CallbackEventListener<typeof tagName$v> | null;
-  accessor onerror: OnErrorEventHandler;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$v]: Image;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$v]: HTMLAttributes<HTMLElement> & ImageJSXProps;
-    }
-  }
-}
-
-declare const tagName$v = 's-image';
-export interface ImageJSXProps
-  extends Partial<ImageProps>,
-    Pick<ImageProps$1, 'id'> {
-  onError?: ((event: CallbackEvent<typeof tagName$v>) => void) | null;
-  onLoad?: ((event: CallbackEvent<typeof tagName$v>) => void) | null;
-}
-
-export type RequiredLinkProps = Required<LinkProps$1>;
-export interface LinkProps
-  extends Pick<
-    LinkProps$1,
-    | 'accessibilityLabel'
-    | 'command'
-    | 'commandFor'
-    | 'download'
-    | 'href'
-    | 'lang'
-    | 'target'
-    | 'tone'
-  > {
-  accessibilityLabel: RequiredLinkProps['accessibilityLabel'];
-  command: RequiredLinkProps['command'];
-  commandFor: RequiredLinkProps['commandFor'];
-  download: RequiredLinkProps['download'];
-  href: RequiredLinkProps['href'];
-  lang: RequiredLinkProps['lang'];
-  target: RequiredLinkProps['target'];
-  tone: Extract<RequiredLinkProps['tone'], 'auto' | 'neutral' | 'critical'>;
-}
-
-declare const Link_base: (abstract new (
-  args_0: RenderImpl,
-) => PreactCustomElement & PreactOverlayControlProps) &
-  Pick<typeof PreactCustomElement, 'observedAttributes' | 'prototype'>;
-declare class Link extends Link_base implements LinkProps {
-  accessor tone: LinkProps['tone'];
-  accessor accessibilityLabel: LinkProps['accessibilityLabel'];
-  accessor href: LinkProps['href'];
-  accessor target: LinkProps['target'];
-  accessor download: LinkProps['download'];
-  accessor lang: LinkProps['lang'];
-  accessor onclick: CallbackEventListener<typeof tagName$u> | null;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$u]: Link;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$u]: HTMLAttributes<HTMLElement> & LinkJSXProps;
-    }
-  }
-}
-
-declare const tagName$u = 's-link';
-export interface LinkJSXProps
-  extends Partial<LinkProps>,
-    Pick<LinkProps$1, 'id' | 'lang'> {
-  onClick?: ((event: CallbackEvent<typeof tagName$u>) => void) | null;
-}
-
-export interface ListItemProps extends ListItemProps$1 {}
-
-declare class ListItem extends PreactCustomElement implements ListItemProps {
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$t]: ListItem;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$t]: HTMLAttributes<HTMLElement> & ListItemJSXProps;
-    }
-  }
-}
-
-declare const tagName$t = 's-list-item';
-export interface ListItemJSXProps
-  extends Partial<ListItemProps>,
-    Pick<ListItemProps$1, 'id'> {}
-
-export type RequiredMoneyFieldProps = Required<MoneyFieldProps$1>;
-export interface MoneyFieldProps
-  extends PreactFieldProps,
-    Pick<RequiredMoneyFieldProps, 'max' | 'min' | 'step' | 'currencyCode'> {}
-
-declare class MoneyField
-  extends PreactFieldElement<MoneyFieldProps['autocomplete']>
-  implements MoneyFieldProps
-{
-  accessor max: MoneyFieldProps['max'];
-  accessor min: MoneyFieldProps['min'];
-  accessor step: MoneyFieldProps['step'];
-  accessor currencyCode: MoneyFieldProps['currencyCode'];
-  get value(): MoneyFieldProps['value'];
-  set value(value: MoneyFieldProps['value']);
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$s]: MoneyField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$s]: HTMLAttributes<HTMLElement> & MoneyFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$s = 's-money-field';
-export interface MoneyFieldJSXProps
-  extends Partial<MoneyFieldProps>,
-    FieldReactProps<typeof tagName$s>,
-    Pick<MoneyFieldProps$1, 'id'> {}
-
-export type NumberFieldProps = PreactFieldProps<
-  Required<NumberFieldProps$1>['autocomplete']
-> &
-  Required<
-    Pick<
-      NumberFieldProps$1,
-      'inputMode' | 'max' | 'min' | 'prefix' | 'step' | 'suffix'
-    >
-  >;
-
-declare class NumberField
-  extends PreactFieldElement<NumberFieldProps['autocomplete']>
-  implements NumberFieldProps
-{
-  get value(): string;
-  set value(value: string);
-  accessor inputMode: NumberFieldProps['inputMode'];
-  accessor step: NumberFieldProps['step'];
-  accessor max: NumberFieldProps['max'];
-  accessor min: NumberFieldProps['min'];
-  accessor prefix: NumberFieldProps['prefix'];
-  accessor suffix: NumberFieldProps['suffix'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$r]: NumberField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$r]: HTMLAttributes<HTMLElement> & NumberFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$r = 's-number-field';
-export interface NumberFieldJSXProps
-  extends Partial<NumberFieldProps>,
-    Pick<NumberFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$r> {}
-
-export interface OptionProps
-  extends Required<
-    Pick<OptionProps$1, 'disabled' | 'value' | 'selected' | 'defaultSelected'>
-  > {}
-
-declare class Option extends PreactCustomElement implements OptionProps {
-  accessor selected: OptionProps['selected'];
-  accessor defaultSelected: OptionProps['defaultSelected'];
-  accessor value: OptionProps['value'];
-  accessor disabled: OptionProps['disabled'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$q]: Option;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$q]: HTMLAttributes<HTMLElement> & OptionJSXProps;
-    }
-  }
-}
-
-declare const tagName$q = 's-option';
-export interface OptionJSXProps extends Partial<OptionProps> {}
-
-export interface OptionGroupProps
-  extends Required<Pick<OptionGroupProps$1, 'disabled' | 'label'>> {}
-
-declare class OptionGroup
-  extends PreactCustomElement
-  implements OptionGroupProps
-{
-  accessor disabled: OptionGroupProps['disabled'];
-  accessor label: OptionGroupProps['label'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$p]: OptionGroup;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$p]: HTMLAttributes<HTMLElement> & OptionGroupJSXProps;
-    }
-  }
-}
-
-declare const tagName$p = 's-option-group';
-export interface OptionGroupJSXProps extends Partial<OptionGroupProps> {}
-
-export interface OrderedListProps extends OrderedListProps$1 {}
-
-declare class OrderedList
-  extends PreactCustomElement
-  implements OrderedListProps
-{
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$o]: OrderedList;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$o]: HTMLAttributes<HTMLElement> & OrderedListJSXProps;
-    }
-  }
-}
-
-declare const tagName$o = 's-ordered-list';
-export interface OrderedListJSXProps
-  extends Partial<OrderedListProps>,
-    Pick<OrderedListProps$1, 'id'> {}
-
-export interface PageProps extends Required<Pick<PageProps$1, 'inlineSize'>> {
-  inlineSize: Extract<PageProps$1['inlineSize'], 'base' | 'large'>;
-}
-
-declare class Page extends PreactCustomElement implements PageProps {
-  accessor inlineSize: PageProps['inlineSize'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$n]: Page;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$n]: HTMLAttributes<HTMLElement> & PageJSXProps;
-    }
-  }
-}
-
-declare const tagName$n = 's-page';
-export interface PageJSXProps extends Partial<PageProps> {
-  aside?: ComponentChild;
-}
-
-export interface ParagraphProps
-  extends Required<
-    Pick<
-      ParagraphProps$1,
-      | 'accessibilityVisibility'
-      | 'fontVariantNumeric'
-      | 'tone'
-      | 'dir'
-      | 'color'
-      | 'lineClamp'
-    >
-  > {
-  color: Extract<ParagraphProps$1['color'], 'base' | 'subdued'>;
-  lineClamp: Extract<ParagraphProps$1['lineClamp'], number>;
-}
-
-declare class Paragraph extends PreactCustomElement implements ParagraphProps {
-  accessor fontVariantNumeric: ParagraphProps['fontVariantNumeric'];
-  accessor lineClamp: ParagraphProps['lineClamp'];
-  accessor tone: Extract<
-    ParagraphProps['tone'],
-    'auto' | 'neutral' | 'info' | 'success' | 'caution' | 'warning' | 'critical'
-  >;
-
-  accessor color: ParagraphProps['color'];
-  accessor dir: ParagraphProps['dir'];
-  accessor accessibilityVisibility: ParagraphProps['accessibilityVisibility'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$m]: Paragraph;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$m]: HTMLAttributes<HTMLElement> & ParagraphJSXProps;
-    }
-  }
-}
-
-declare const tagName$m = 's-paragraph';
-export interface ParagraphJSXProps
-  extends Partial<ParagraphProps>,
-    Pick<ParagraphProps$1, 'id'> {}
-
-export type PasswordFieldProps = PreactFieldProps<
-  Required<PasswordFieldProps$1>['autocomplete']
-> &
-  Required<
-    Pick<
-      PasswordFieldProps$1,
-      | 'defaultValue'
-      | 'details'
-      | 'disabled'
-      | 'error'
-      | 'labelAccessibilityVisibility'
-      | 'minLength'
-      | 'maxLength'
-      | 'label'
-      | 'name'
-      | 'placeholder'
-      | 'readOnly'
-      | 'required'
-      | 'value'
-    >
-  >;
-
-declare class PasswordField
-  extends PreactFieldElement<PasswordFieldProps['autocomplete']>
-  implements PasswordFieldProps
-{
-  accessor maxLength: PasswordFieldProps['maxLength'];
-  accessor minLength: PasswordFieldProps['minLength'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$l]: PasswordField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$l]: HTMLAttributes<HTMLElement> & PasswordFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$l = 's-password-field';
-export interface PasswordFieldJSXProps
-  extends Partial<PasswordFieldProps>,
-    Pick<PasswordFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$l> {}
-
-export type SearchFieldProps = PreactFieldProps<
-  Required<TextFieldProps$1>['autocomplete']
-> &
-  Required<
-    Pick<
-      TextFieldProps$1,
-      | 'defaultValue'
-      | 'details'
-      | 'disabled'
-      | 'error'
-      | 'labelAccessibilityVisibility'
-      | 'minLength'
-      | 'maxLength'
-      | 'label'
-      | 'name'
-      | 'placeholder'
-      | 'readOnly'
-      | 'required'
-      | 'value'
-    >
-  >;
-
-declare class SearchField
-  extends PreactFieldElement<SearchFieldProps['autocomplete']>
-  implements SearchFieldProps
-{
-  accessor maxLength: SearchFieldProps['maxLength'];
-  accessor minLength: SearchFieldProps['minLength'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$k]: SearchField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$k]: HTMLAttributes<HTMLElement> & SearchFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$k = 's-search-field';
-export interface SearchFieldJSXProps
-  extends Partial<SearchFieldProps>,
-    Pick<TextFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$k> {}
-
-export type RequiredSectionProps = Required<SectionProps$1>;
-export interface SectionProps
-  extends Pick<
-    RequiredSectionProps,
-    'accessibilityLabel' | 'heading' | 'padding'
-  > {
-  accessibilityLabel: RequiredSectionProps['accessibilityLabel'];
-  heading: RequiredSectionProps['heading'];
-  padding: RequiredSectionProps['padding'];
-}
-
-declare class Section extends PreactCustomElement implements SectionProps {
-  constructor();
-  connectedCallback(): void;
-  accessor accessibilityLabel: SectionProps['accessibilityLabel'];
-  accessor heading: SectionProps['heading'];
-  accessor padding: SectionProps['padding'];
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$j]: Section;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$j]: HTMLAttributes<HTMLElement> & SectionJSXProps;
-    }
-  }
-}
-
-declare const tagName$j = 's-section';
-export interface SectionJSXProps
-  extends Partial<SectionProps>,
-    Pick<SectionProps$1, 'id'> {}
-
-export type SelectProps = PreactInputProps &
-  Required<
-    Pick<
-      SelectProps$1,
-      | 'details'
-      | 'disabled'
-      | 'error'
-      | 'label'
-      | 'name'
-      | 'placeholder'
-      | 'required'
-      | 'value'
-      | 'icon'
-      | 'labelAccessibilityVisibility'
-    >
-  >;
-
-declare const usedFirstOptionSymbol: unique symbol;
-declare const hasInitialValueSymbol: unique symbol;
-
-declare class Select extends PreactInputElement implements SelectProps {
-  accessor icon: SelectProps['icon'];
-  accessor details: SelectProps['details'];
-  accessor error: SelectProps['error'];
-  accessor label: SelectProps['label'];
-  accessor placeholder: SelectProps['placeholder'];
-  accessor required: SelectProps['required'];
-  accessor labelAccessibilityVisibility: SelectProps['labelAccessibilityVisibility'];
-  connectedCallback(): void;
-  disconnectedCallback(): void;
-  constructor();
-  /**
-   * used to determine if no value or defaultValue was set, in which case the first non-disabled option was used
-   *
-   * this is important because we need to use the placeholder in these situations, even though the first value will be submitted as part of the form
-   */
-  [usedFirstOptionSymbol]: boolean;
-  [hasInitialValueSymbol]: boolean;
-  get value(): string;
-  set value(value: string);
-  formResetCallback(): void;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$i]: Select;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$i]: HTMLAttributes<HTMLElement> & SelectJSXProps;
-    }
-  }
-}
-
-declare const tagName$i = 's-select';
-export interface SelectJSXProps extends Partial<SelectProps> {
-  onChange?: (event: CallbackEvent<typeof tagName$i>) => void;
-  onInput?: (event: CallbackEvent<typeof tagName$i>) => void;
-  onBlur?: (event: CallbackEvent<typeof tagName$i>) => void;
-  onFocus?: (event: CallbackEvent<typeof tagName$i>) => void;
-}
-
-export interface SpinnerProps
-  extends Required<Pick<SpinnerProps$1, 'accessibilityLabel'>> {
-  /**
-   * The size of the component.
-   *
-   * @default 'base'
-   */
-  size: Extract<SpinnerProps$1['size'], 'large' | 'large-100' | 'base'>;
-}
-
-declare class Spinner extends PreactCustomElement implements SpinnerProps {
-  accessor accessibilityLabel: string;
-  accessor size: SpinnerProps['size'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$h]: Spinner;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$h]: Omit<HTMLAttributes<HTMLElement>, 'size'> & SpinnerJSXProps;
-    }
-  }
-}
-
-declare const tagName$h = 's-spinner';
-export interface SpinnerJSXProps
-  extends Partial<SpinnerProps>,
-    Pick<SpinnerProps$1, 'id'> {}
-
-export type AlignedStackProps = Required<StackProps$1>;
-export interface StackProps
-  extends BoxProps,
-    Pick<
-      Required<AlignedStackProps>,
-      'justifyContent' | 'alignItems' | 'alignContent'
-    > {
-  justifyContent: AlignedStackProps['justifyContent'];
-  alignItems: AlignedStackProps['alignItems'];
-  alignContent: AlignedStackProps['alignContent'];
-  gap: MakeResponsive<AlignedStackProps['gap']>;
-  rowGap: MakeResponsive<AlignedStackProps['rowGap']>;
-  columnGap: MakeResponsive<AlignedStackProps['columnGap']>;
-  direction: MakeResponsive<AlignedStackProps['direction']>;
-}
-
-declare class Stack extends BoxElement implements StackProps {
-  constructor();
-  accessor direction: StackProps['direction'];
-  accessor justifyContent: StackProps['justifyContent'];
-  accessor alignItems: StackProps['alignItems'];
-  accessor alignContent: StackProps['alignContent'];
-  accessor gap: StackProps['gap'];
-  accessor rowGap: StackProps['rowGap'];
-  accessor columnGap: StackProps['columnGap'];
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$g]: Stack;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$g]: HTMLAttributes<HTMLElement> & StackJSXProps;
-    }
-  }
-}
-
-declare const tagName$g = 's-stack';
-export interface StackJSXProps
-  extends Partial<StackProps>,
-    Pick<StackProps$1, 'id'> {}
-
-export interface SwitchProps
-  extends PreactCheckboxProps,
-    Required<Pick<SwitchProps$1, 'labelAccessibilityVisibility'>> {}
-
-declare class Switch extends PreactCheckboxElement implements SwitchProps {
-  accessor labelAccessibilityVisibility: SwitchProps['labelAccessibilityVisibility'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$f]: Switch;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$f]: HTMLAttributes<HTMLElement> & SwitchJSXProps;
-    }
-  }
-}
-
-declare const tagName$f = 's-switch';
-export interface SwitchJSXProps
-  extends Partial<SwitchProps>,
-    Pick<SwitchProps$1, 'id'> {
-  onChange?: ((event: CallbackEvent<typeof tagName$f>) => void) | null;
-  onInput?: ((event: CallbackEvent<typeof tagName$f>) => void) | null;
-}
-
-export interface TableProps
-  extends Required<
-    Pick<
-      TableProps$1,
-      'loading' | 'paginate' | 'hasPreviousPage' | 'hasNextPage' | 'variant'
-    >
-  > {
-  variant: Extract<TableProps$1['variant'], 'list' | 'auto'>;
-}
-
-export interface TableHeaderProps extends Pick<TableHeaderProps$1, 'listSlot'> {
-  listSlot: Extract<
-    TableHeaderProps$1['listSlot'],
-    'primary' | 'secondary' | 'labeled' | 'kicker' | 'inline'
-  >;
-}
-
-declare class AddedContext<T> extends EventTarget {
-  constructor(defaultValue: T);
-  get value(): T;
-  set value(value: T);
-}
-
-declare const actualTableVariantSymbol: unique symbol;
-declare const tableHeadersSharedDataSymbol: unique symbol;
-export type ActualTableVariant = 'table' | 'list';
-
-declare class Table extends PreactCustomElement implements TableProps {
-  accessor variant: TableProps['variant'];
-  accessor loading: TableProps['loading'];
-  accessor paginate: TableProps['paginate'];
-  accessor hasPreviousPage: TableProps['hasPreviousPage'];
-  accessor hasNextPage: TableProps['hasNextPage'];
-  accessor onpreviouspage: CallbackEventListener<typeof tagName$e> | null;
-  accessor onnextpage: CallbackEventListener<typeof tagName$e> | null;
-  /**
-   * The actual table variant, which is either 'table' or 'list'.
-   */
-  [actualTableVariantSymbol]: AddedContext<ActualTableVariant>;
-  [tableHeadersSharedDataSymbol]: AddedContext<
-    {
-      listSlot: TableHeaderProps['listSlot'];
-      textContent: string;
-    }[]
-  >;
-
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$e]: Table;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$e]: HTMLAttributes<HTMLElement> & TableJSXProps;
-    }
-  }
-}
-
-declare const tagName$e = 's-table';
-export interface TableJSXProps
-  extends Partial<TableProps>,
-    Pick<TableProps$1, 'id' | 'onNextPage' | 'onPreviousPage'> {
-  filters?: ComponentChild;
-}
-
-export interface TableBodyProps extends TableBodyProps$1 {}
-
-declare class TableBody extends PreactCustomElement implements TableBodyProps {
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$d]: TableBody;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$d]: HTMLAttributes<HTMLElement> & TableBodyJSXProps;
-    }
-  }
-}
-
-declare const tagName$d = 's-table-body';
-export interface TableBodyJSXProps
-  extends Partial<TableBodyProps>,
-    Pick<TableBodyProps$1, 'id'> {}
-
-export interface TableCellProps extends TableCellProps$1 {}
-
-declare class TableCell extends PreactCustomElement implements TableCellProps {
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$c]: TableCell;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$c]: HTMLAttributes<HTMLElement> & TableCellJSXProps;
-    }
-  }
-}
-
-declare const tagName$c = 's-table-cell';
-export interface TableCellJSXProps
-  extends Partial<TableCellProps>,
-    Pick<TableCellProps$1, 'id'> {}
-
-declare class TableHeader
-  extends PreactCustomElement
-  implements TableHeaderProps
-{
-  accessor listSlot: TableHeaderProps['listSlot'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$b]: TableHeader;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$b]: HTMLAttributes<HTMLElement> & TableHeaderJSXProps;
-    }
-  }
-}
-
-declare const tagName$b = 's-table-header';
-export interface TableHeaderJSXProps
-  extends Partial<TableHeaderProps>,
-    Pick<TableHeaderProps$1, 'id'> {}
-
-export interface TableHeaderRowProps extends TableHeaderRowProps$1 {}
-
-declare class TableHeaderRow
-  extends PreactCustomElement
-  implements TableHeaderRowProps
-{
-  constructor();
-  connectedCallback(): void;
-  disconnectedCallback(): void;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$a]: TableHeaderRow;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$a]: HTMLAttributes<HTMLElement> & TableHeaderRowJSXProps;
-    }
-  }
-}
-
-declare const tagName$a = 's-table-header-row';
-export interface TableHeaderRowJSXProps
-  extends Partial<TableHeaderRowProps>,
-    Pick<TableHeaderRowProps$1, 'id'> {}
-
-export interface TableRowProps extends TableRowProps$1 {}
-
-declare class TableRow extends PreactCustomElement implements TableRowProps {
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$9]: TableRow;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$9]: HTMLAttributes<HTMLElement> & TableRowJSXProps;
-    }
-  }
-}
-
-declare const tagName$9 = 's-table-row';
-export interface TableRowJSXProps
-  extends Partial<TableRowProps>,
-    Pick<TableRowProps$1, 'id'> {}
-
-export interface TextProps
-  extends Required<
-    Pick<
-      TextProps$1,
-      | 'display'
-      | 'accessibilityVisibility'
-      | 'dir'
-      | 'color'
-      | 'type'
-      | 'tone'
-      | 'fontVariantNumeric'
-    >
-  > {
-  color: Extract<TextProps$1['color'], 'base' | 'subdued'>;
-  type: Extract<
-    TextProps$1['type'],
-    'address' | 'redundant' | 'strong' | 'generic'
-  >;
-  tone: Extract<
-    TextProps$1['tone'],
-    'auto' | 'info' | 'success' | 'warning' | 'critical' | 'neutral' | 'caution'
-  >;
-  fontVariantNumeric: Extract<
-    TextProps$1['fontVariantNumeric'],
-    'auto' | 'normal' | 'tabular-nums'
-  >;
-}
-
-declare class Text extends PreactCustomElement implements TextProps {
-  accessor display: TextProps['display'];
-  accessor fontVariantNumeric: TextProps['fontVariantNumeric'];
-  accessor color: TextProps['color'];
-  accessor tone: TextProps['tone'];
-  accessor type: TextProps['type'];
-  accessor dir: TextProps['dir'];
-  accessor accessibilityVisibility: TextProps['accessibilityVisibility'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$8]: Text;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$8]: HTMLAttributes<HTMLElement> & TextJSXProps;
-    }
-  }
-}
-
-declare const tagName$8 = 's-text';
-export interface TextJSXProps
-  extends Partial<TextProps>,
-    Pick<TextProps$1, 'id'> {}
-
-export type TextAreaProps = PreactFieldProps<
-  Required<TextAreaProps$1>['autocomplete']
-> &
-  Required<Pick<TextAreaProps$1, 'maxLength' | 'minLength' | 'rows'>>;
-
-declare class TextArea
-  extends PreactFieldElement<TextAreaProps['autocomplete']>
-  implements TextAreaProps
-{
-  accessor maxLength: TextAreaProps['maxLength'];
-  accessor minLength: TextAreaProps['minLength'];
-  accessor rows: TextAreaProps['rows'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$7]: TextArea;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$7]: HTMLAttributes<HTMLElement> & TextAreaJSXProps;
-    }
-  }
-}
-
-declare const tagName$7 = 's-text-area';
-export interface TextAreaJSXProps
-  extends Partial<TextAreaProps>,
-    Pick<TextAreaProps$1, 'id'>,
-    FieldReactProps<typeof tagName$7> {}
-
-export type TextFieldProps = PreactFieldProps<
-  Required<TextFieldProps$1>['autocomplete']
-> &
-  Required<
-    Pick<
-      TextFieldProps$1,
-      'icon' | 'maxLength' | 'minLength' | 'prefix' | 'suffix'
-    >
-  >;
-
-declare class TextField
-  extends PreactFieldElement<TextFieldProps['autocomplete']>
-  implements TextFieldProps
-{
-  accessor icon: TextFieldProps['icon'];
-  accessor maxLength: TextFieldProps['maxLength'];
-  accessor minLength: TextFieldProps['minLength'];
-  accessor prefix: TextFieldProps['prefix'];
-  accessor suffix: TextFieldProps['suffix'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$6]: TextField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$6]: HTMLAttributes<HTMLElement> & TextFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$6 = 's-text-field';
-export interface TextFieldJSXProps
-  extends Partial<Omit<TextFieldProps, 'accessory'>>,
-    Pick<TextFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$6> {
-  accessory?: ComponentChild;
-}
-
-export type URLFieldProps = PreactFieldProps<
-  Required<URLFieldProps$1>['autocomplete']
-> &
-  Required<Pick<URLFieldProps$1, 'maxLength' | 'minLength'>>;
-
-declare class URLField
-  extends PreactFieldElement<URLFieldProps['autocomplete']>
-  implements URLFieldProps
-{
-  accessor autocomplete: URLFieldProps['autocomplete'];
-  accessor maxLength: URLFieldProps['maxLength'];
-  accessor minLength: URLFieldProps['minLength'];
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$5]: URLField;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$5]: HTMLAttributes<HTMLElement> & URLFieldJSXProps;
-    }
-  }
-}
-
-declare const tagName$5 = 's-url-field';
-export interface URLFieldJSXProps
-  extends Partial<Omit<URLFieldProps, 'accessory'>>,
-    Pick<URLFieldProps$1, 'id'>,
-    FieldReactProps<typeof tagName$5> {}
-
-export interface UnorderedListProps extends UnorderedListProps$1 {}
-
-declare class UnorderedList
-  extends PreactCustomElement
-  implements UnorderedListProps
-{
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$4]: UnorderedList;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$4]: HTMLAttributes<HTMLElement> & UnorderedListJSXProps;
-    }
-  }
-}
-
-declare const tagName$4 = 's-unordered-list';
-export interface UnorderedListJSXProps
-  extends Partial<UnorderedListProps>,
-    Pick<UnorderedListProps$1, 'id'> {}
-
-export interface AdminActionProps
-  extends Pick<AdminActionProps$1, 'heading' | 'loading'> {}
-
-declare const tagName$3 = 's-admin-action';
-export interface AdminActionJSXProps
-  extends Partial<AdminActionProps>,
-    Pick<AdminActionProps$1, 'id'> {}
-
-declare class AdminAction
-  extends PreactCustomElement
-  implements AdminActionProps
-{
-  heading: string;
-  loading: boolean;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$3]: AdminAction;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$3]: HTMLAttributes<HTMLElement> & AdminActionJSXProps;
-    }
-  }
-}
-
-export interface AdminBlockProps
-  extends Pick<AdminBlockProps$1, 'heading' | 'collapsedSummary'> {}
-
-declare const tagName$2 = 's-admin-block';
-export interface AdminBlockJSXProps
-  extends Partial<AdminBlockProps>,
-    Pick<AdminBlockProps$1, 'id'> {}
-
-declare class AdminBlock
-  extends PreactCustomElement
-  implements AdminBlockProps
-{
-  heading: string;
-  collapsedSummary: string;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$2]: AdminBlock;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$2]: HTMLAttributes<HTMLElement> & AdminBlockJSXProps;
-    }
-  }
-}
-
-export interface AdminPrintActionProps
-  extends Pick<AdminPrintActionProps$1, 'src'> {}
-
-declare const tagName$1 = 's-admin-print-action';
-export interface AdminPrintActionJSXProps
-  extends Partial<AdminPrintActionProps>,
-    Pick<AdminPrintActionProps$1, 'id'> {}
-
-declare class AdminPrintAction
-  extends PreactCustomElement
-  implements AdminPrintActionProps
-{
-  src: string;
-  constructor();
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$1]: AdminPrintAction;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$1]: HTMLAttributes<HTMLElement> & AdminPrintActionJSXProps;
-    }
-  }
-}
-
-declare const tagName = 's-form';
-export interface FormProps {}
-declare class Form extends PreactCustomElement implements FormProps {
-  constructor();
-  accessor onsubmit: ((event: ExtendableEvent) => void) | null;
-  accessor onreset: ((event: ExtendableEvent) => void) | null;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName]: Form;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName]: HTMLAttributes<HTMLElement> & FormJSXProps;
-    }
-  }
-}
-
-export interface FormJSXProps extends Partial<FormProps> {
-  id?: string;
-  /**
-   * A callback that is run when the form is submitted.
-   */
-  onSubmit?: ((event: ExtendableEvent) => void) | null;
-  /**
-   * A callback that is run when the form is reset.
-   */
-  onReset?: ((event: ExtendableEvent) => void) | null;
-}
-
-export {
-  AdminAction,
-  type AdminActionJSXProps,
-  AdminBlock,
-  type AdminBlockJSXProps,
-  AdminPrintAction,
-  type AdminPrintActionJSXProps,
-  Badge,
-  type BadgeJSXProps,
-  Banner,
-  type BannerJSXProps,
-  Box,
-  type BoxJSXProps,
-  Button,
-  type ButtonJSXProps,
-  Checkbox,
-  type CheckboxJSXProps,
-  Choice,
-  type ChoiceJSXProps,
-  ChoiceList,
-  type ChoiceListJSXProps,
-  Clickable,
-  type ClickableJSXProps,
-  Divider,
-  type DividerJSXProps,
-  EmailField,
-  type EmailFieldJSXProps,
-  Form,
-  type FormJSXProps,
-  Grid,
-  GridItem,
-  type GridItemJSXProps,
-  type GridJSXProps,
-  Heading,
-  type HeadingJSXProps,
-  Icon,
-  type IconJSXProps,
-  Image,
-  type ImageJSXProps,
-  Link,
-  type LinkJSXProps,
-  ListItem,
-  type ListItemJSXProps,
-  MoneyField,
-  type MoneyFieldJSXProps,
-  NumberField,
-  type NumberFieldJSXProps,
-  Option,
-  OptionGroup,
-  type OptionGroupJSXProps,
-  type OptionJSXProps,
-  OrderedList,
-  type OrderedListJSXProps,
-  Page,
-  type PageJSXProps,
-  Paragraph,
-  type ParagraphJSXProps,
-  PasswordField,
-  type PasswordFieldJSXProps,
-  SearchField,
-  type SearchFieldJSXProps,
-  Section,
-  type SectionJSXProps,
-  Select,
-  type SelectJSXProps,
-  Spinner,
-  type SpinnerJSXProps,
-  Stack,
-  type StackJSXProps,
-  Switch,
-  type SwitchJSXProps,
-  Table,
-  TableBody,
-  type TableBodyJSXProps,
-  TableCell,
-  type TableCellJSXProps,
-  TableHeader,
-  type TableHeaderJSXProps,
-  TableHeaderRow,
-  type TableHeaderRowJSXProps,
-  type TableJSXProps,
-  TableRow,
-  type TableRowJSXProps,
-  Text,
-  TextArea,
-  type TextAreaJSXProps,
-  TextField,
-  type TextFieldJSXProps,
-  type TextJSXProps,
-  URLField,
-  type URLFieldJSXProps,
-  UnorderedList,
-  type UnorderedListJSXProps,
-};
