--- packages/ui-extensions/src/surfaces/admin/extension.ts
@@ -1,7 +1,0 @@
-import {createExtensionRegistrationFunction} from '../../extension';
-
-import type {ExtensionTargets} from './extension-targets';
-
-export const extension =
-  createExtensionRegistrationFunction<ExtensionTargets>();
-export const extend = extension;
