--- packages/ui-extensions/src/surfaces/admin/globals.ts
@@ -1,16 +1,0 @@
-import type {ExtensionTargets} from './extension-targets';
-
-export interface ShopifyGlobal {
-  extend<ExtensionTarget extends keyof ExtensionTargets>(
-    target: ExtensionTarget,
-    extend: ExtensionTargets[ExtensionTarget],
-    version?: string,
-  ): void;
-}
-
-declare global {
-  // conflicts with build/ts/globals.d.ts
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  const shopify: ShopifyGlobal;
-}
